### 2024-06-12 #6879 @editorialbot
[PRE REVIEW]: scores: A Python package for verifying and evaluating models and predictions with xarray and pandas
**Submitting author:** @tennlee (<a href="http://orcid.org/0009-0008-2024-1967">Tennessee Leeuwenburg</a>)
**Repository:** https://github.com/nci/scores/
**Branch with paper.md** (empty if default branch): 528-joss-paper-submission
**Version:** 0.9.0
