### 2024-06-12 #6879 @crvernon
:wave:  @tennlee - Thanks for your submission to JOSS.  I'll take this one on myself.  As soon as I get you two reviewers signed up, we will kick this over to full review.  Please feel free to offer a few suggestions for reviewers by providing their GitHub handle without the `@` so we don't notify them prematurely.  Thanks!
