### 2024-06-13 0b670ff
[feature] utility functions to help with dimension ordering/alignment
