### 2024-06-13 1673b9d
Update to README including citation information for arXiv preprint
