### 2024-06-13 b154f77
Update README.md
Update to include arXiv citation information

Signed-off-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
