### 2024-06-13 322c5a6
[stubs] Add stubs for helpers to check input arrays in block bootstrapping
