+++ src/scores/emerging/block_bootstrap/helpers.py
@@ -13,6 +13,14 @@ from typing import TypeVar, Tuple
 T = TypeVar("T")
 
 
+def unique_or_error(xs: list[T]):
+    """
+    Checks if input list has unique elements, otherwise raise an error.
+    """
+    if len(set(xs)) != len(xs):
+        raise ValueError("List elements must be unique.")
+
+
 def partial_linear_order_by_ref(xs: list[T], xs_ref: list[T]) -> Tuple[list[T], list[T]]:
     """
     Sorts elements in ``xs`` in the order that they appear in the ``xs_ref``.
@@ -22,32 +30,39 @@ def partial_linear_order_by_ref(xs: list[T], xs_ref: list[T]) -> Tuple[list[T], 
 
     Args:
         xs: list of arbitrary type, with elements that can be ordered
-        xs_ref: reference list, same type as xs
+        xs_ref: reference list, same type as xs, must have unique elements
 
     Returns:
         A partitioned list where,
-            - The first list contains a linearly ordered copy of ``xs``
-            - The second contains elements in `xs` that could not be found in
+            - The first list contains a partial linearly ordered copy of ``xs``
+            - The second contains elements in ``xs`` that could not be found in
               ``xs_ref``
     Raises:
         ValueError: If list is not linearly ordered, i.e. has gaps.
 
-    In this case, a linearly ordered list, ``xs`` is ordered in reference to a
-    list ``xs_ref`` such that:
+    In this case, a partially linearly ordered list, ``xs`` is ordered in
+    reference to a list ``xs_ref`` such that:
 
     .. code-block::
-        xs[i] == xs_ref[i] for all i < len(xs)
+        if xs[i] == xs_ref[i] for all i < len(xs)
+
+        then xs is partially linearly ordered.
 
         where,
 
         the comparision (<) reflects the order the elements appear in xs_ref:
         xs_ref[0] < xs_ref[1] < xs_ref[2] ...
     """
-
+    # empty lists and lists with 1 element are ordered
     if len(xs) <= 1:
         return xs
 
+    unique_or_error(xs_ref)
+
+    # sort xs in the same order as xs_ref
     xs_sorted = [x for x in xs_ref if x in xs]
+
+    # check for partial linear ordering as per description in docstring
     is_linord = all([xs_sorted[i] == x for i, x in enumerate(xs_ref[0 : len(xs_sorted)])])
 
     if not is_linord:
@@ -58,6 +73,7 @@ def partial_linear_order_by_ref(xs: list[T], xs_ref: list[T]) -> Tuple[list[T], 
             f" {xs_ref}, the following are gaps in the data: {xs_gaps}."
         )
 
+    # also return elements missing in xs_ref for parent stack to handle
     xs_missing = list(set(xs) - set(xs_sorted))
 
     return (xs_sorted, xs_missing)
@@ -68,10 +84,38 @@ def reorder_dims(
     dims: list[str],
     auto_order_missing: bool = True,
 ) -> xr.DataArray:
-    # (ordered dims, `arr.dims` not present in `dims`)
+    """
+    Reorders the dimensions of all arrays, with respect to a reference list
+    of dimensions (``dims``), so that they are compatible.
+
+    If ``dims`` does not contain a dimension in ``arr``, it cannot be
+    ordered, and will cause an error to be raised. Optionally, if the
+    ``auto_order_missing`` flag is ``True`` (default), then missing
+    dimensions will be sent to the back (innermost axes) and ordered
+    alphebatically so as to maintain consistency between all arrays.
+
+    Args:
+
+        arrs: data array reorder
+        dims: unique list of reference dimensions that define the expected
+            order.
+        auto_order_missing: optionally order dimensinons in ``arrs`` that
[…]
