### 2024-06-13 ceb0fdc
Improved docstrings for BinaryContingencyManager and for BasicContingencyManager
