### 2024-06-13 eebc4ea
Docstring tweaks
