### 2024-06-13 ea1ba2a
Change 'forecast' and 'observed' to 'fcst' and 'obs' in a small number of places for greater consistency
