### 2024-06-13 fe4a747
Reduce line length on mathjax lines Address a small number of other pylint concerns
