### 2024-06-13 13cd043
Update src/scores/categorical/contingency_impl.py
Co-authored-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
Signed-off-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
