### 2024-06-13 ce9208a
Change 'forecast' and 'observed' to 'fcst' and 'obs' in a small (#537)
* Change 'forecast' and 'observed' to 'fcst' and 'obs' in a small number of places for greater consistency
* Rename forecast_events to fcst_events and observed_events to obs_events
