### 2024-06-13 306e4ed
Fix multiline formula in dot point
