### 2024-06-13 9af69fd
Reduce line length on mathjax lines (#539)
* Reduce line length on mathjax lines
* Address a small number of other pylint concerns
