+++ src/scores/continuous/standard_impl.py
@@ -210,7 +210,8 @@ def correlation(
     Calculates the Pearson's correlation coefficient between two xarray DataArrays
 
     .. math::
-        \\rho = \\frac{\\sum_{i=1}^{n}{(x_i - \\bar{x})(y_i - \\bar{y})}}{\\sqrt{\\sum_{i=1}^{n}{(x_i-\\bar{x})^2}\\sum_{i=1}^{n}{(y_i - \\bar{y})^2}}}
+        \\rho = \\frac{\\sum_{i=1}^{n}{(x_i - \\bar{x})(y_i - \\bar{y})}}
+        {\\sqrt{\\sum_{i=1}^{n}{(x_i-\\bar{x})^2}\\sum_{i=1}^{n}{(y_i - \\bar{y})^2}}}
 
     where:
         - :math:`\\rho` = Pearson's correlation coefficient
