+++ src/scores/categorical/contingency_impl.py
@@ -18,6 +18,8 @@ Scores supports complex, weighted, multi-dimensional data, including in continge
 Users can supply their own event operators to the top-level module functions.
 """
 
+# pylint: disable=too-many-lines
+
 import operator
 from abc import ABC, abstractmethod
 from typing import Optional
@@ -367,7 +369,8 @@ class BasicContingencyManager:  # pylint: disable=too-many-public-methods
             xr.DataArray: An xarray object containing the probability of false detection
 
         .. math::
-            \\text{probability of false detection} = \\frac{\\text{false positives}}{\\text{true negatives} + \\text{false positives}}
+            \\text{probability of false detection} = \\frac{\\text{false positives}}{\\text{true negatives} +
+            \\text{false positives}}
 
         Notes:
             - Range: 0 to 1.  Perfect score: 0.
@@ -390,7 +393,8 @@ class BasicContingencyManager:  # pylint: disable=too-many-public-methods
             xr.DataArray: An xarray object containing the success ratio
 
         .. math::
-            \\text{success ratio} = \\frac{\\text{true positives}}{\\text{true positives} + \\text{false positives}}
+            \\text{success ratio} = \\frac{\\text{true positives}}{\\text{true positives} +
+            \\text{false positives}}
 
         Notes:
             - Range: 0 to 1.  Perfect score: 1.
@@ -413,7 +417,8 @@ class BasicContingencyManager:  # pylint: disable=too-many-public-methods
             xr.DataArray: An xarray object containing the threat score
 
         .. math::
-            \\text{threat score} = \\frac{\\text{true positives}}{\\text{true positives} + \\text{false positives} + \\text{false negatives}}
+            \\text{threat score} = \\frac{\\text{true positives}}{\\text{true positives} +
+            \\text{false positives} + \\text{false negatives}}
 
         Notes:
             - Range: 0 to 1, 0 indicates no skill. Perfect score: 1.
@@ -438,7 +443,8 @@ class BasicContingencyManager:  # pylint: disable=too-many-public-methods
             xr.DataArray: An xarray object containing the critical success index
 
         .. math::
-            \\text{threat score} = \\frac{\\text{true positives}}{\\text{true positives} + \\text{false positives} + \\text{false negatives}}
+            \\text{threat score} = \\frac{\\text{true positives}}{\\text{true positives} +
+            \\text{false positives} + \\text{false negatives}}
 
         Notes:
             - Range: 0 to 1, 0 indicates no skill. Perfect score: 1.
@@ -462,7 +468,9 @@ class BasicContingencyManager:  # pylint: disable=too-many-public-methods
             xr.DataArray: An xarray object containing the Peirce Skill Score
 
         .. math::
-            \\text{Peirce skill score} = \\frac{\\text{true positives}}{\\text{true positives} + \\text{false negatives}} - \\frac{\\text{false positives}}{\\text{false positives} + \\text{true negatives}}
+            \\text{Peirce skill score} = \\frac{\\text{true positives}}{\\text{true positives} + 
+            \\text{false negatives}} - \\frac{\\text{false positives}}{\\text{false positives} + 
+            \\text{true negatives}}
 
         Notes:
             - Range: -1 to 1, 0 indicates no skill. Perfect score: 1.
@@ -492,7 +500,9 @@ class BasicContingencyManager:  # pylint: disable=too-many-public-methods
             xr.DataArray: An xarray object containing the true skill statistic
 
         .. math::
[…]
