### 2024-06-14 8037c50
Update src/scores/categorical/contingency_impl.py
Co-authored-by: Nikeeth Ramanathan <nikeeth.ramanathan@gmail.com>
Signed-off-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
