### 2024-06-14 b82c6bd
531 documentation uplift for contingency manager classes (#535)
* Improved docstrings for BinaryContingencyManager and for BasicContingencyManager

---------

Signed-off-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
Co-authored-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
Co-authored-by: Nikeeth Ramanathan <nikeeth.ramanathan@gmail.com>
