+++ src/scores/categorical/contingency_impl.py
@@ -35,14 +35,25 @@ DEFAULT_PRECISION = 8
 
 class BasicContingencyManager:  # pylint: disable=too-many-public-methods
     """
-    A BasicContingencyManager is produced when a BinaryContingencyManager is transformed.
+    See https://scores.readthedocs.io/en/stable/tutorials/Binary_Contingency_Scores.html for
+    a detailed walkthrough showing the use of this class in practice.
 
-    A basic contingency table is built only from the event counts, losing the connection
-    to the actual event tables in their full dimensionality.
+    A BasicContingencyManager object provides the scoring functions which are calculated
+    from a contingency table. A BasicContingencyManager can be efficiently and repeatedly
+    queried for a wide variety of scores.
 
-    The event count data is much smaller than the full event tables, particularly when
-    considering very large data sets like Numerical Weather Prediction (NWP) data, which
-    could be terabytes to petabytes in size.
+    A BasicContingencyManager is produced when a :py:class:`BinaryContingencyManager` is
+    transformed. It is also possible to create a BasicContingencyManager from event counts or
+    a contingency table, although this not a common user requirement.
+
+    A contingency table is built only from event counts, losing the connection
+    to the actual event tables in their full dimensionality. The event count data is much
+    smaller than the full event tables, particularly when considering very large data sets
+    like Numerical Weather Prediction (NWP) data, which could be terabytes to petabytes in
+    size.
+
+    By contrast, A :py:class:`BinaryContingencyManager` retains the full event data, which provides
+    some more flexbility but may reduce efficiency.
     """
 
     def __init__(self, counts: dict):
@@ -82,13 +93,63 @@ class BasicContingencyManager:  # pylint: disable=too-many-public-methods
 
     def get_counts(self) -> dict:
         """
-        Return the contingency table counts (tp, fp, tn, fn)
+
+        Returns:
+            dict: A dictionary of the contingency table counts. Values are xr.DataArray instances.
+            The keys are:
+
+            - tp_count for true positive count
+            - tn_count for true negative count,
+            - fp_count for false positive count
+            - fn_count for false negative count
+            - total_count for the total number of events.
+
+        Here is an example of what may be returned:
+
+        .. code-block :: python
+
+            {'tp_count': <xarray.DataArray ()> Size: 8B array(5.),
+             'tn_count': <xarray.DataArray ()> Size: 8B array(11.),
+             'fp_count': <xarray.DataArray ()> Size: 8B array(1.),
+             'fn_count': <xarray.DataArray ()> Size: 8B array(1.),
+             'total_count': <xarray.DataArray ()> Size: 8B array(18.)}
+
         """
         return self.counts
 
     def get_table(self) -> xr.DataArray:
         """
-        Return the contingency table as an xarray object
+        Returns:
+            xr.DataArray: The contingency table as an xarray object. Contains
+            a coordinate dimension called 'contingency'. Valid coordinates
+            for this dimenstion are:
+
+            - tp_count for true positive count
+            - tn_count for true negative count
+            - fp_count for false positive count
+            - fn_count for false negative count
+            - total_count for the total number of events.
+
+        Here is an example of what may be returned:
+
+        .. code-block :: python
+
+            array([ 5., 11.,  1.,  1., 18.])
+
+            Coordinates:
+
+                contingency
+                (contingency)
+                <U11
+                'tp_count' ... 'total_count'
+
+            Indexes:
+
+                contingency
+                PandasIndex
+
+            Attributes: (0)
+
         """
         return self.xr_table
 
@@ -1031,27 +1092,38 @@ class BasicContingencyManager:  # pylint: disable=too-many-public-methods
 
[…]
