+++ src/scores/__init__.py
@@ -13,7 +13,7 @@ import scores.processing
 import scores.sample_data
 import scores.stats.statistical_tests  # noqa: F401
 
-__version__ = "1.0.0"
+__version__ = "0.9.1"
 
 __all__ = [
     "scores.categorical",
