### 2024-06-14 #6879 @tennlee
Hi there. We just released version 0.9.1 of `scores` as we had some important updates. I'm not sure of the implications for the review process and so I thought I should mention it here.
