+++ docs/included.md
@@ -15,15 +15,15 @@
   - [Mean Error (WWRP/WGNE Joint Working Group on Forecast Verification Research)](https://www.cawcr.gov.au/projects/verification/#meanerror)
 * - Consistent Expectile Score
   - [API](api.md#scores.continuous.consistent_expectile_score)
-  - [Tutorial](project:./tutorials/Consistent_Scores.html)
+  - [Tutorial](project:./tutorials/Consistent_Scores.md)
   - [Gneiting (2011)](https://doi.org/10.1198/jasa.2011.r101388), [Taggart (2022)](https://doi.org/10.1002/qj.4206)
 * - Consistent Huber Score
   - [API](api.md#scores.continuous.consistent_huber_score)
-  - [Tutorial](project:./tutorials/Consistent_Scores.html)
+  - [Tutorial](project:./tutorials/Consistent_Scores.md)
   - [Taggart (2022a)](https://doi.org/10.1002/qj.4206), [Taggart (2022b)](https://doi.org/10.1214/21-EJS1957)
 * - Consistent Quantile Score
   - [API](api.md#cores.continuous.consistent_quantile_score)
-  - [Tutorial](project:./tutorials/Consistent_Scores.html)
+  - [Tutorial](project:./tutorials/Consistent_Scores.md)
   - [Gneiting (2011)](https://doi.org/10.1198/jasa.2011.r101388), [Taggart (2022)](https://doi.org/10.1002/qj.4206)
 * - Flip-Flop Index
   - 
