### 2024-06-14 848ca67
[wip] stubs to expand implementation to support xarray
