+++ src/scores/emerging/block_bootstrap/block_bootstrap.py
@@ -22,7 +22,110 @@ from scores.emerging.block_bootstrap.axis_info import AxisInfo, make_axis_info
 from scores.emerging.block_bootstrap.methods import FitBlocksMethod
 
 
-def sample_axis_block_indices(
+# TODO: A lot of the same variables are common to many of the functions, causing a lot of
+# repetition there should probably be two classes (namespaces) to hold state:
+# 1. atomic computation: using numpy/numba to read/write the actual block sample
+# 2. public interface: for handling xaray inputs and wider options as well as parallelization
+#    options.
+
+def block_bootstrap(
+    arrs: list[xr.DataArray],
+    bootstrap_dims: list[str],
+    block_sizes: list[int],
+    iterations: int,
+    fit_blocks_method: FitBlocksMethod = FitBlocksMethod.PARTIAL,
+    cyclic: bool = True,
+    auto_order_missing: bool = True,
+) -> list[xr.DataArray]:
+
+    # validate inputs
+    if iterations <= 0:
+        ValueError("`iterations` must be greater than 0.")
+
+    if len(block_sizes) == len(bootstrap_dims):
+        ValueError("`block_sizes` must be the same size as `bootstrap_dims`.")
+
+
+    raise NotImplementedError("`block_bootstrap` is currently a stub.")
+
+    # reorder dimensions to align across all arrays
+    arrs_reordered = reorder_all_arr_dims(arrs, bootstrap_dims, auto_order_missing)
+
+    # TODO:
+    # - collect all array dims
+    # - make axis info for each dimension
+    # - add dimension name for axis info to make it uniquely identifiable
+    # - create axis block indices for dimension `[ (axis_info, axis_block_sample_indices) ]`
+    # - if dimension does not exist in bootstrap_dims, set bootstrap = False when creating index,
+    #   so that parts of the algorithm can use it.
+
+    arrs_bootstrapped = []
+
+    for arr in arrs:
+        # TODO:
+        # - get axis subset for array
+        # - perform ufunc on `_construct_block_bootstrap_array` with core dims as bootstrap_dims
+        # - expand to `iterations`
+        # - append to result list.
+        pass
+
+    return arrs_bootstrapped
+
+
+def _construct_block_bootstrap_array(
+    input_arr: npt.NDArray,
+    block_sizes: list[int],
+    cyclic: bool = True,
+    fit_blocks_method: FitBlocksMethod = FitBlocksMethod.PARTIAL,
+) -> npt.NDArray:
+    """
+    takes a numpy array and performs block resampling for 1 iteration,
+
+    note that the output array size is not guarenteed to be the same shape,
+    depending on FitBlocksMethod
+    """
+    assert len(block_sizes) == len(input_arr.shape)
+
+    # construct axis info based on block sizes and input array
+    ax_info = make_axis_info(input_arr, block_sizes, fit_blocks_method)
+
+    # get sample block indices for each axis
+    ax_block_indices = _sample_axis_block_indices(ax_info, cyclic)
+
+    # re-fetch block-sizes (since internal functions may update this, e.g. based on dask chunks)
+    ax_blk_sizes = [axi.block_size for axi in ax_info]
+    ax_num_blks = [axi.num_blocks for axi in ax_info]
+    ax_len_out = [axi.length_out for axi in ax_info]
+
+    # construct output array, based on axis info
+    output_arr = np.empty(ax_len_out)
+
+    # dummy array for block index looping
+    dummy_block_idx_arr = np.empty(ax_num_blks)
+    num_blocks_iter = np.nditer(dummy_block_idx_arr, flags=["multi_index"], op_flags=["readonly"])
+
+    with num_blocks_iter:
+        for _ in num_blocks_iter:
[…]
