### 2024-06-14 ebf8ac4
341 potential unused function (#543)
* Add crps_step_threshold_weight to the public API
