+++ src/scores/probability/__init__.py
@@ -9,6 +9,7 @@ from scores.probability.crps_impl import (
     crps_cdf,
     crps_cdf_brier_decomposition,
     crps_for_ensemble,
+    crps_step_threshold_weight,
 )
 from scores.probability.roc_impl import roc_curve_data
 from scores.processing.isoreg_impl import isotonic_fit
@@ -23,4 +24,5 @@ __all__ = [
     "crps_for_ensemble",
     "roc_curve_data",
     "isotonic_fit",
+    "crps_step_threshold_weight",
 ]
