+++ docs/included.md
@@ -137,7 +137,15 @@
   -
     [API](api.md#scores.probability.adjust_fcst_for_crps)
   - 
-    [Tutorial](project:./tutorials/CRPS_for_CDFs.md)
+    &mdash;
+  - 
+    &mdash;
+* -  
+    - CRPS step threshold weights
+  -
+    [API](api.md#scores.crps_step_threshold_weight)
+  - 
+    &mdash;
   - 
     &mdash;
 * -  
