### 2024-06-14 c6893f2
Post-review adjustments to the index of included scores for step threshold weights
