+++ docs/included.md
@@ -22,7 +22,7 @@
   - [Tutorial](project:./tutorials/Consistent_Scores.md)
   - [Taggart (2022a)](https://doi.org/10.1002/qj.4206); [Taggart (2022b)](https://doi.org/10.1214/21-EJS1957)
 * - Consistent Quantile Score
-  - [API](api.md#cores.continuous.consistent_quantile_score)
+  - [API](api.md#scores.continuous.consistent_quantile_score)
   - [Tutorial](project:./tutorials/Consistent_Scores.md)
   - [Gneiting (2011)](https://doi.org/10.1198/jasa.2011.r101388); [Taggart (2022)](https://doi.org/10.1002/qj.4206)
 * - Flip-Flop Index
