### 2024-06-18 nci/scores#546 @savente93
[JOSS] Lisence is not correctly specified
This issue was files as part of my JOSS review for https://github.com/openjournals/joss-reviews/issues/6889.

While the project is licensed under the Apache License, Version 2.0, which is an OSI approved license, the license is only linked to, causing automated systems such as GitHub and the editorial bot to not recognize the license. Both in the interest of the authors work, the digital providence and automated license checking I'd like to see this rectified.
