### 2024-06-18 nci/scores#547 @savente93
[JOSS] Consider tracking coverage 
This issue is opened as part of the JOSS review process of https://github.com/openjournals/joss-reviews/issues/6889

The authors of this package admirably require 100% code coverage. While this means that tracking code coverage over time, with something like `codecov.io` (or any other available solution) might not be very useful to the authors, having this information available is very useful for potential users trying to evaluate the project. Given that the authors make the effort to keep 100% test coverage, they might as well gain the credit from it (as well as easily have it verified). I would suggest the authors add some code coverage tracking solution and perhaps a badge displaying it, as is quite common. Though this is by no means required, and they are free to dismiss this concern if they wish.
