### 2024-06-18 nci/scores#548 @savente93
[JOSS] Number scheme is not explained 
this issue was created as part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/6889

From the commit/PR history as well as the test workfows, it seems that there is some numbering scheme in place (e.g. branches are often named `xxx-branch-name` to the point there it will influence whether tests get run. However this scheme is not documented anywhere that I could find. I would like to see the documentation at least explain:
1. what the numbers mean (even if they don't mean anything beyond sorting)
2. how to generate a next number for contributors
3. any potential implications of choosing (or not choosing) a number.
