### 2024-06-18 nci/scores#549 @savente93
[JOSS] JOSS statement of need should be expanded 
this issue was created as part of my JOSS review https://github.com/openjournals/joss-reviews/issues/6889

While the statement of need does state what the software is used for, but not what previously unmet requirement it fulfills. There is a lot of information about this in the rest of the paper, but is also somewhat inconsistent in my opinion. Was `scores` made to develop new metrics? if so has that happened? or was there before this a lack of verification methods/error metrics? Or was it made to provide an open version of Jive?

Continuing on the topic of prior work or (possibly) lack of other packages with verification methods, given that the authors mention the application of machine learning or economics in the first paragraph of the paper I would have expected comparisons to either `scipy` or `scikit-learn` since those are very well known packages that also implement a lot of these metrics. From this I deduce that the need was actually for verification/error metrics in _N dimensional data_ but this is not specified. While I have not gone through the tutorials yet, so this may change as I continue my review, as a data scientist I would not see the need for this package, so unless the application to machine learning or economics can be demonstrated more clearly I would honestly remove it.

On the same note, I would suggest the authors elaborate terms like "forecast" or "model" a bit more. This might seem silly, but from my work with hydrologists I know that data scientists and earth sciences can use these terms in very different ways, so to avoid confusion here it would be good to quickly explain the terms.

Most of the information is there, but I think it should be shuffled around a bit. I hope this has been constructive enough, feel free to ask questions you might have :)
