### 2024-06-18 #6889 @savente93
I've gone through a lot of the auxiliary stuff and  made some (hopefully useful) remarks, I'll continue my review of the functionality etc. etc. at a later point.
