### 2024-06-19 nci/scores#549 @tennlee
Thanks for these comments. As I understand it, we're encouraged to have these discussions as we go, rather than waiting for the review to conclude. We'd be very happy to amend our statement of need to address these comments. We agree the paper needs to make its points clearly. It's very helpful to get a new perspective.

I wanted to provide a brief clarifying reply, so you don't have any unanswered questions as you proceed with the rest of the review.

 - Scores uses xarray, which is the most commonly used package to support the labelled, n-dimensional data utilised by the meteorological and Earth system community. Verification/stats/metrics packages that are used commonly in data science often only support 1-D data, tabular data, or unlabelled n-dimensional data (see "Data Handling" in statement of need).
 - Scores handles dimensionality preferably to most other meteorological packages, and weighting (such as latitude-weighting) preferably to many general data science packages
 - Scores provides some novel metrics, developed by our team, which aren't in any other package.
 - Commonly-used scores are also provided for completeness and consistency, so `scores` can be used standalone, and also to meet frequent user requests for their inclusion.
 - It would typically take several packages to achieve most of what scores provides - and even then would not provide all the key benefits of scores.

We will take some more time to consider how to improve the paper to incorporate your feedback and respond further when we have done so.

We have also been cautious in specifically saying what other packages can't do, or can't do as effectively, both to avoid implied criticism of others' work and also for clarity. We have avoided 'stating the negative' about what other packages don't do. No other package meets all the elements of our statement of need.
