### 2024-06-19 nci/scores#548 @tennlee
Thanks. There is not a lot of formality to the numbers - they are generated by default by GitHub for some branch creation workflows based on issues, and we (or I at least) find them somewhat helpful. I'll add some explanation to the contributor's guide, but nobody should feel obliged to use them, and you don't need to try to follow any scheme during the review process.

The numbering is not intended to filter whether tests are run - we expect the tests to run on every branch.

The branch "228-documentation-testing-branch" is used differently to everything else in `scores`. This is a working branch which is used as a space to see how readthedocs will render a change. As such, it will change frequently and never be merged directly. Only some developers are comfortable building the docs locally, and also there are still potential differences between a local build and the eventual rendering in readthedocs.

I find the branch numbers very helpful in distinguishing similarly-named or ambiguously-named branches, but it's a feature intended for the human eye and no automated processes rely on or use the branch numbering. It is also somewhat historical as most of the core team was originally using branches on the main repository, however as we move to a fork-based workflow, the numbers are less  relevant and no longer created by default.
