### 2024-06-19 nci/scores#546 @tennlee
Hopefully I have now fixed this. I used the GitHub interface for the change, which means the develop branch is now updated and the repository itself now shows "Apache 2.0" for the license where it previously privded "Show license". Hopefully this means it is now done in standard fashion. If you are happy to confirm this looks right to you also, I would appreciate it.

The change will flow through to the main branch when we next prepare a release. If I have understood correctly, most tools/integrations rely on the license specified in the default branch. In our case that is the develop branch so I think it is all in order now.
