### 2024-06-19 #6889 @tennlee
Running this to see if the automated check is now happy with the changes to the license specification.
