+++ .binder/requirements.txt
@@ -297,7 +297,7 @@ rpds-py==0.18.0
     # via
     #   jsonschema
     #   referencing
-scikit-learn==1.4.1.post1
+scikit-learn==1.5.0
     # via scores (pyproject.toml)
 scipy==1.12.0
     # via
