### 2024-06-19 9b7d738
Update 'np.NaN' to 'np.nan' in two places in one test file np.NaN has been deprecated and should no longer be used
