### 2024-06-19 #6889 @tennlee
I suspect the bot did not pick up the license from the default branch. I have copied the new license file into the same branch as contains paper.md.  Running the check again.
