### 2024-06-19 #6889 @tennlee
Note for reviewers - GitHub's security scanner, dependabot, noted two dependencies which needed updating. These both pertained to the "binder" integration and I have updated the versions accordingly on our default branch. The changes are unrelated to the main dependencies of scores and are unlikely to affect anyone not using binder, but I am happy to update branch "528-joss-paper-submission" if that is wanted, just let me know if you would like that.
