### 2024-06-19 nci/scores#546 @tennlee
A quick note - the JOSS editorialbot is happy now and shows as green tick for the license.
