### 2024-06-19 nci/scores#549 @savente93
>  As I understand it, we're encouraged to have these discussions as we go, rather than waiting for the review to conclude.

That is also how I understand it yes, so feel free to interact with me or push changes while I'm busy, I'll then try and look at the changes when I have time :)

As for the clarifying replies, these are all really good points, and I would consider adding these in some integrated way to the paper a really good way to resolve my comments. For me this is mostly about highlighting the work you have done already

> We have also been cautious in specifically saying what other packages can't do, or can't do as effectively, both to avoid implied criticism of others' work and also for clarity. We have avoided 'stating the negative' about what other packages don't do. No other package meets all the elements of our statement of need.

That is a sensible position to have. I am still on the opinion that pointing out the "weaknesses" in prior work and how this compares to your own is part of the academic process and helps newcomers to your package make an informed decision on whether your package is the right one for them, and that this can be done without being demeaning or overly negative (though that is certainly hard). I don't want to push you to make claims you're not comfortable making, but if possible I'd suggest to give this another pass. I don't consider it a hard requirement for the publication though.
