### 2024-06-19 nci/scores#552 @savente93
[JOSS] Mention binder availability in README
This issue was opened as part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/6889

The documentation of the tutorials offer binder integration which is quite nice. However, this is not mentioned in the README, meaning some users will not see it. I think this is a really nice feature that it is worth mentioning in the README. In my experience it lowers the barrier to entry, but only if people know it is there. For example, I'd already made an environment before I realized the binder integration was an option. You might consider either adding a binder badge to the README, or mentioning the option when you introduce the tutorials.
