### 2024-06-19 nci/scores#553 @savente93
[JOSS] Add a end-to-end example
This issue was filed as part of my JOSS review:  https://github.com/openjournals/joss-reviews/issues/6889

This one I've seen on your road map a couple of times so I think you're already aware of it, but I'm filing this issue  anyway just to highlight the (in my eyes) importance of it. Having a real world data example, that not just demonstrates it's real world use, but the whole end-to-end process including analyzing and drawing conclusions is really quite invaluable for news users. The tutorials on the individual metrics are great if those are the metrics you are interested in using, but an end-to-end example, is a real foothold for users trying to assess and contextualize your package, and also for less experienced users trying to gain experience. Encountering an end-to-end example, not just the "interesting" bits always significantly increases my option of a project, even if the actual exercise or "interesting part" is relatively small and boring. It helps to show the real effort required in using a package.

Again, this is just a suggestion, so I don't consider this a requirement for accepting the publication, but I would strongly encourage it. The authors already have a robust workflow for making these, so I assume that adding one of these is a relatively high ROI activity.
