### 2024-06-19 nci/scores#554 @savente93
[JOSS] Fix warnings in test suite and examples
This issue was filed as part of my JOSS review https://github.com/openjournals/joss-reviews/issues/6889

There are a few warnings generated by both the tutorials and the test suite. They are very superficial warnings, such as deprecation warnings, so it is not a critical issue, but it is good practice to keep your examples and test suite warning free if possible. I realize this is a bit of a pet peeve of mine, so the submission acceptance is not conditional on fixing this, but I would recommend it, as it's an easy win and gives the package a much more polished impression for new comers.
