### 2024-06-19 #6889 @savente93
@tennlee @crvernon I am done with my initial review. I've made issues on the repository with some suggestions for improvements. I think the only issue I'd like to see addressed for giving scores a passing mark is  https://github.com/nci/scores/issues/549 all the other ones are optional.

As a general remark, the documentation of `scores` is excellent. I do consider <400 commits over 2 (ish) years a somewhat immature package, but it definitely meets the criteria. If it fulfilled a need I had I would feel comfortable giving this package a serious try.

I'll be leaving things off here for now but I should be available for comments and questions whenever necessary.
