### 2024-06-19 38fa51c
Add Python 3.12 to the list of tested versions in github actions
