### 2024-06-19 nci/scores#553 @nicholasloveday
I'll let @tennlee respond in more detail, but some simple end-to-end examples exist in the github repository for [this paper](https://arxiv.org/abs/2404.18429 ) on Jive.
For example, see https://github.com/nicholasloveday/jive_paper/blob/main/src/dailypop_calibration_fig.ipynb which is used to produce figure 8 in the paper.

The use cases in that repository are fairly simple, because it's a fairly high level paper that we submitted to BAMS.

Some more complex examples of using scores can be found in [this paper](https://journals.ametsoc.org/view/journals/wefo/aop/WAF-D-23-0201.1/WAF-D-23-0201.1.xml?tab_body=pdf), but that code isn't publicly available.
