### 2024-06-20 3f4bae9
increase test cov for consistent scores tests
