+++ .github/workflows/python-app.yml
@@ -37,3 +37,7 @@ jobs:
     - name: Test tutorial notebooks with pytest
       run: |
         pytest --nbmake tutorials/
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v4.0.1
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
