### 2024-06-20 f05344b
added consistent scores to readme
