### 2024-06-20 c8601f2
Add code coverage badge
