### 2024-06-20 86394cc
Update python-app.yml
Signed-off-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
