+++ .github/workflows/python-app.yml
@@ -4,7 +4,8 @@
 name: Unit Tests
 
 on:
-  push
+  pull_request:
+    branches: ["develop", "main", "[0-9]+-*"]
 
 permissions:
   contents: read
