### 2024-06-20 a4cadec
Add Badges to the README for CodeQL, code coverage, and binder link (#555)
Set up inital badges and codecov.io integration.
A second PR will be needed to update the badge URL after CodeCov is triggered to actually create the coverage report on its side.
