### 2024-06-20 e928ae6
Remove code coverage badge for the time being
