### 2024-06-20 5490842
[wip] support and align dimensions for multiple xarray inputs
