### 2024-06-20 nci/scores#553 @savente93
That is fine, as a short term solution you could consider mentioning that somewhere in the tutorial gallery as well, though every jump away from your repository will lose you interest, so in a slightly longer term I would still consider your own.
