### 2024-06-20 nci/scores#554 @tennlee
Thanks for your feedback and I also prefer to handle all warnings. In this case, I will need to do some investigation in how to do this before I can proceed. As such it might take me some time. I thought I should provide a response in the meantime so it's clear we intend to pick up the issue.

My results are that the Python 3.12 tests are showing no warnings, whereas I see warnings on the other Python versions. I don't know the best way to handle version-specific warning handling across the test suite. The warnings all relate to code in dependencies so all we could do is catch them as expected for that version.

The warnings in the tutorials bother me more than the warnings in the tests, since they are presented more directly to scientific users who may be less able to quickly identify what to disregard and what to worry about. I will probably raise a separate issue to track the notebooks so that I can chunk up the work more easily.
