### 2024-06-20 nci/scores#552 @tennlee
I have now added a badge to the README showing the binder integration, but haven't yet considered how (or if) to include it elsewhere in the README. I just wanted to provide an update here since there is no linked pull request for that change.
