+++ .github/workflows/python-app.yml
@@ -37,7 +37,6 @@ jobs:
     - name: Test tutorial notebooks with pytest
       run: |
         pytest --nbmake tutorials/
-    - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v4.0.1
-      with:
-        token: ${{ secrets.CODECOV_TOKEN }}
+    - name: Coveralls GitHub Action
+      uses: coverallsapp/github-action@v2.3.0
+  
