### 2024-06-20 3f3d1e0
Add coveralls badge Hopefully, commit now triggers on every push and on every pull request, covering all types of contribution to the repo
