+++ .github/workflows/python-app.yml
@@ -4,7 +4,7 @@
 name: Unit Tests
 
 on:
-  push
+  [push, pull_request]
 
 permissions:
   contents: read
