### 2024-06-20 ec49d52
Move coverage badge location slightly
