### 2024-06-20 nci/scores#547 @tennlee
I have added a code coverage badge, thanks for the prompt! I believe this resolves the issue, but I will leave the issue open for a few days in case there is more to discuss.
