### 2024-06-20 nci/scores#552 @tennlee
On reflection, I think adding the badge to the README seems sufficient. I don't want to further complicate the documentation with another topic. I'm happy to add some descriptive text to the tutorial gallery if you think it's worthwhile. Due to the need to download data to get the most out of interactive use of scores, the binder experience is somewhat limited, and I think a simple approach is suitable. As a result, I'm going to close this issue, but please do feel free to re-open it or leave further comment and I'll happily revisit it.
