### 2024-06-20 nci/scores#547 @tennlee
A further update - I do think this is now resolved, so I am going to close the issue now. Please don't hesitate to re-open it or leave a comment if there is anything else that needs to be done, and I will happily revisit it.
