### 2024-06-20 fa5f28f
Add commentary on branch numbering to the contributing guide
