### 2024-06-21 ecbd064
Respond to merge request feedback and add additional details
