### 2024-06-21 e616cda
Add commentary on branch numbering to the contributing guide (#557)
* Add commentary on branch numbering to the contributing guide
