### 2024-06-21 839ea4c
Update coveralls action to utilise parallel build
