### 2024-06-21 1d4c071
Add referencing style info and some other edits for clarity
