### 2024-06-21 9a8dda7
Incorporate feedback from Beth, and add WeatherBench 2 and some other minor changes
