### 2024-06-21 9f2abe2
Add link to WeatherBench 2 datasets (so users can more easily find datasets)
