### 2024-06-21 af42d71
Add further description of point-based observations
