+++ docs/data.md
@@ -39,7 +39,7 @@ Archived datasets are available for:
 
 #### Point-Based Data
 
-Point-based observations are shared routinely between countries for the purposes of weather modelling.
+Point-based observations (e.g. from weather stations or buoys) are shared routinely between countries for the purposes of weather modelling.
 
 The NOAA Integrated Surface Database (ISD) provides hourly point-based (*in-situ*) weather station data globally. It is a good starting point for understanding how to work with point-based data. For more information about the NOAA ISD see [https://www.ncei.noaa.gov/products/land-based-station/integrated-surface-database](https://www.ncei.noaa.gov/products/land-based-station/integrated-surface-database).
 
