### 2024-06-21 nci/scores#554 @savente93
> My results are that the Python 3.12 tests are showing no warnings, whereas I see warnings on the other Python versions. I don't know the best way to handle version-specific warning handling across the test suite. The warnings all relate to code in dependencies so all we could do is catch them as expected for that version.

If they are warnings that you are aware of and know don't impact you there are several things you can do:
1. you could add filter warnings for pytest in your `pyproject.toml`  (see https://docs.pytest.org/en/7.1.x/how-to/capture-warnings.html#deprecationwarning-and-pendingdeprecationwarning)
2. You could add filter warnings to your `__init__.py` (see https://docs.python.org/3/library/warnings.html#overriding-the-default-filter )
3. Not by any means necessary, but if you're feeling up to it you could go to upstream and file either a PR or an issue to fix the warning

These are the options I usually go for, I think which you choose is mostly up to preference.
