### 2024-06-21 3494dda
[wip] axis collection; psuedo tests for axis checks; docstring for main bootstrapping method
