+++ src/scores/emerging/block_bootstrap/_test_block_bootstrap_TO_REMOVE.py
@@ -5,18 +5,24 @@ import pprint
 import numpy as np
 import xarray as xr
 
-from scores.emerging.block_bootstrap.axis_info import AxisInfo, make_axis_info
+from scores.emerging.block_bootstrap.axis_info import (
+    AxisInfo,
+    make_axis_info,
+    make_axis_info_collection,
+)
 from scores.emerging.block_bootstrap.block_bootstrap import (
     _construct_block_bootstrap_array,
 )
 from scores.emerging.block_bootstrap.helpers import (
     partial_linear_order_by_ref,
+    reorder_all_arr_dims,
     reorder_dims,
 )
 from scores.emerging.block_bootstrap.methods import FitBlocksMethod
 
 
 def _test_numpy_blk_bootstrap_single_iter():
+    print("\nBOOTSTRAPPING (single iteration, single array)...")
     # generate test data
     axis_len = [13, 10, 8, 7]
     block_sizes = [4, 3, 2, 3]
@@ -25,7 +31,7 @@ def _test_numpy_blk_bootstrap_single_iter():
     # random dataset with integers so its easy to visualize
     input_arr = rng.integers(low=0, high=10, size=axis_len)
 
-    print("--- axis info ---")
+    print("\n--- axis info ---")
     axis_info = make_axis_info(input_arr, block_sizes, method)
     pprint.pp(axis_info)
 
@@ -34,7 +40,6 @@ def _test_numpy_blk_bootstrap_single_iter():
     res = np.histogram(input_arr, bins=5)
     pprint.pp(res)
 
-    print("\nBOOTSTRAPPING...")
     (output_arr, block_sample_idx) = _construct_block_bootstrap_array(
         input_arr,
         block_sizes,
@@ -52,7 +57,106 @@ def _test_numpy_blk_bootstrap_single_iter():
     pprint.pp(res)
 
 
-def _test_order_dims():
+def _test_make_axis_info_collection():
+    print("\nAXIS COLLECTION...")
+
+    arr_test_1 = np.random.rand(10, 10, 12, 7, 5)
+    da_1 = xr.DataArray(
+        data=arr_test_1,
+        dims=["x", "y", "time", "lead_time", "height"],
+    )
+    arr_test_2 = np.random.rand(10, 10, 7, 5)
+    da_2 = xr.DataArray(
+        data=arr_test_2,
+        dims=["y", "x", "lead_time", "height"],
+    )
+    arr_test_3 = np.random.rand(10, 10, 7)
+    da_3 = xr.DataArray(
+        data=arr_test_3,
+        dims=["x", "y", "lead_time"],
+    )
+    axi_collection = make_axis_info_collection(
+        [da_1, da_2, da_3],
+        ["lead_time", "x", "y"],
+        [2, 4, 5],
+    )
+
+    print("\n--- axis collection ---")
+    pprint.pp(axi_collection)
+
+    try:
+        make_axis_info_collection(
+            [da_1, da_2, da_3],
+            ["lead_time", "x", "y"],
+            [2, 4, 12],
+        )
+    except AssertionError:
+        print(f"\n--- failure due to invalid block size ---")
+        print(f"Caught expected assertion error")
+
+    try:
+        arr_test_fail = np.random.rand(10, 11, 7)
+        da_fail = xr.DataArray(
+            data=arr_test_fail,
+            dims=["x", "y", "lead_time"],
+        )
+        make_axis_info_collection(
+            [da_1, da_2, da_fail],
+            ["lead_time", "x", "y"],
+            [2, 4, 5],
+        )
+    except ValueError as e:
[…]
