### 2024-06-22 684a275
Potential update to joss paper (#18)
Potential update to JOSS paper
