### 2024-06-22 a8968b1
Adjust sentence in statement of need
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
