### 2024-06-22 #6889 @tennlee
Hi all. I have rebased the branch "528-joss-paper-submission" which contains the paper.md file against "develop" which is our default and development branch. This fixes some test failures arising from shifting versions in the dependencies, and brings in various documentation updates (including some of those recommended by the reviewers). I was unclear whether to keep the 528 branch up to date or not, but in light of failing tests occurring, it seemed the most practical choice. I can revert this situation if requested. Thanks.
