### 2024-06-23 553ff45
Apply suggestions from code review
Co-authored-by: Nikeeth Ramanathan <nikeeth.ramanathan@gmail.com>
Co-authored-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
