### 2024-06-23 050ddcf
Adding contingency table metrics to categorical section of table
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
