### 2024-06-23 fce0a58
add brackets around contingency table metrics reference in table
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
