### 2024-06-23 f718b55
Update sentence in first paragraph, update sentence in table
