### 2024-06-23 c865438
Several minor updates to README
