### 2024-06-24 d745355
Update two sentences in first paragraph
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
