### 2024-06-24 1027be5
Update two sentences in README
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
