### 2024-06-24 533dcd0
Minor updates to README (#564)
* Several minor updates to README

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>

---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
