### 2024-06-24 3d13c56
Potential update to joss paper (#18) (#561)
* Draft paper for submission to JOSS
For editing history, see branch 51-joss-paper-draft
This branch will track the submission process

* Adjust in statement of need
* Adding contingency table metrics to categorical section of table
