### 2024-06-24 #6889 @tennlee
@savente93 @calebweinreb The authors have updated the paper (paper.md) on branch 528-joss-paper-submission so it is now ready for further review. We have also made some updates to the documentation (see scores.readthedocs.io/en/latest)
