### 2024-06-24 nci/scores#549 @tennlee
The authors have updated the paper (paper.md) on branch 528-joss-paper-submission so it is now ready for further review. We have also made some updates to the documentation (see https://scores.readthedocs.io/en/latest)
