### 2024-06-24 642b724
Adds a Key Benefits page to docs
