### 2024-06-24 4f6e520
Fix error in FIRM dot point
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
