### 2024-06-24 nci/scores#554 @tennlee
I believe I have now addressed all of the issues in the tutorials, at least for Python 3.12. This is the version used to build the documentation and the error no longer shows in readthedocs latest. It is also the version developers/users should get in new environments.
