### 2024-06-24 nci/scores#554 @savente93
The tutorials indeed seem to be warning free so that is very nice. Though I'm still seeing some deprecatio warnings in the test suite:
```
DeprecationWarning: `trapz` is deprecated. Use `trapezoid` instead, or one of the numerical integration functions in `scipy.integrate`.
```
see pytest section at https://github.com/nci/scores/actions/runs/9645482074/job/26599796094
