### 2024-06-25 54aed49
Update and rename key_benefits.md to key_features.md
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
