### 2024-06-25 b22181d
Update index.md to include key_features
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
