+++ docs/index.md
@@ -12,7 +12,7 @@
 :maxdepth: 2
 
 self
-key_benefits
+key_features
 installation
 included
 api
