### 2024-06-25 nci/scores#554 @tennlee
Thanks. I will work through the test suites next and report back.
