### 2024-06-25 a5ec802
Update dot points relating to included metrics
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
