### 2024-06-25 nci/scores#549 @savente93
I'm happy with the updated submission, I think it highlights the strengths of `scores` a lot better. Well done :)
