### 2024-06-25 #6889 @savente93
I've reviewed the updated paper, and I'm happy with what has been provided, so as far as I'm concerned the `scores` submission is ready to be accepted. There are a few things I've mentioned yet, but those are not necessary for acceptance and might take a bit longer, so I'll just leave that up to the authors (I believe they've already made significant progress on them). @crvernon I've completed my checklist, is anything else needed from my side?
