### 2024-06-25 c345c1e
(ready for review) Add a Key Features page to docs (#567)
* Adds a Key Features page to docs

---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
