### 2024-06-25 c0f9ca1
added consistent Huber score equation
