+++ src/scores/continuous/consistent_impl.py
@@ -24,9 +24,28 @@ def consistent_expectile_score(
     weights: Optional[xr.DataArray] = None,
 ) -> xr.DataArray:
     """
-    Returns the score using a scoring function that is consistent for the
+    Calculates the score using a scoring function that is consistent for the
     alpha-expectile functional, based on a supplied convex function phi.
     See Geniting (2011), or Equation (10) from Taggart (2021).
+
+    .. math::
+
+        S(x, y) =
+        \\begin{cases}
+        (1 - \\alpha)(\\phi(y) - \\phi(x) - \\phi'(x)(y-x)), & y < x \\\\
+        \\alpha(\\phi(y) - \\phi(x) - \\phi'(x)(y-x)), & x \\leq y
+        \\end{cases}
+
+    where
+
+        - :math:`x` is the forecast
+        - :math:`y` is the observation
+        - :math:`\\alpha` is the expectile level
+        - :math:`\\phi` is a convex function of a single variable
+        - :math:`\\phi'` is the subderivative of :math:`\\phi`
+        - :math:`S(x,y)` is the score.
+
+    Note that if :math:`\\phi` is differentiable then `\\phi'` is its derivative.
 
     Args:
         fcst: array of forecast values.
@@ -59,24 +78,7 @@ def consistent_expectile_score(
         ValueError: if `alpha` is not strictly between 0 and 1.
 
     Note:
-        .. math::
-
-            S(x, y) =
-            \\begin{cases}
-            (1 - \\alpha)(\\phi(y) - \\phi(x) - \\phi'(x)(y-x)), & y < x \\\\
-            \\alpha(\\phi(y) - \\phi(x) - \\phi'(x)(y-x)), & x \\leq y
-            \\end{cases}
-
-        where
-
-            - :math:`x` is the forecast
-            - :math:`y` is the observation
-            - :math:`\\alpha` is the expectile level
-            - :math:`\\phi` is a convex function of a single variable
-            - :math:`\\phi'` is the subderivative of :math:`\\phi`
-            - :math:`S(x,y)` is the score.
-
-        Note that if :math:`\\phi` is differentiable then `\\phi'` is its derivative.
+        The score is negatively oriented with :math:`0 \\leq S(x,y) < \\infty `.
 
     References:
         -   Gneiting, T. (2011). Making and Evaluating Point Forecasts. Journal of the 
@@ -89,8 +91,7 @@ def consistent_expectile_score(
     """
     check_alpha(alpha)
 
-    if preserve_dims or reduce_dims:
-        reduce_dims = gather_dimensions(fcst.dims, obs.dims, reduce_dims=reduce_dims, preserve_dims=preserve_dims)
+    reduce_dims = gather_dimensions(fcst.dims, obs.dims, reduce_dims=reduce_dims, preserve_dims=preserve_dims)
 
     score_overfcst = (1 - alpha) * (phi(obs) - phi(fcst) - phi_prime(fcst) * (obs - fcst))
     score_underfcst = alpha * (phi(obs) - phi(fcst) - phi_prime(fcst) * (obs - fcst))
@@ -113,9 +114,23 @@ def consistent_huber_score(
     weights: Optional[xr.DataArray] = None,
 ) -> xr.DataArray:
     """
-    Score that is consistent for the Huber mean functional with tuning parameter `tuning_param`,
-    based on convex function phi. See Taggart (2022), or Equation (11) from Taggart (2021).
-    See Taggart (2021), end of Section 3.4, for the standard formula.
+    Calculates the score that is consistent for the Huber mean functional with tuning
+    parameter `tuning_param`, based on convex function phi. See Taggart (2022), or
[…]
