### 2024-06-26 9d68697
add reccomendation to work through tutorial
