### 2024-06-26 476d9fe
fix equations in docs
