### 2024-06-26 4365a93
hopefully fixed equations in docs
