### 2024-06-26 53a673a
fix score range equation
