### 2024-06-26 ed1a2be
Introduce API for consistent scoring rules
Introduces consistent scoring rules to the codebase

There are mathematically specialised scores which will mainly be internally used by an upcoming API for threshold-weighted scores, but may also be of interest to end-users in some cases.

---------

Signed-off-by: Nicholas Loveday <48701367+nicholasloveday@users.noreply.github.com>
Co-authored-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
