+++ src/scores/continuous/consistent_impl.py
@@ -1,0 +1,284 @@
+"""
+Implementation of scoring functions that are consistent for
+single-valued forecasts targeting quantiles, expectiles or Huber functionals.
+"""
+
+from typing import Callable, Optional
+
+import xarray as xr
+
+from scores.functions import apply_weights
+from scores.typing import FlexibleDimensionTypes
+from scores.utils import gather_dimensions
+
+
+def consistent_expectile_score(
+    fcst: xr.DataArray,
+    obs: xr.DataArray,
+    alpha: float,
+    phi: Callable[[xr.DataArray], xr.DataArray],
+    phi_prime: Callable[[xr.DataArray], xr.DataArray],
+    *,  # Force keywords arguments to be keyword-only
+    reduce_dims: Optional[FlexibleDimensionTypes] = None,
+    preserve_dims: Optional[FlexibleDimensionTypes] = None,
+    weights: Optional[xr.DataArray] = None,
+) -> xr.DataArray:
+    """
+    Calculates the score using a scoring function that is consistent for the
+    alpha-expectile functional, based on a supplied convex function phi.
+    See Geniting (2011), or Equation (10) from Taggart (2022).
+
+    .. math::
+
+        S(x, y) =
+        \\begin{cases}
+        (1 - \\alpha)(\\phi(y) - \\phi(x) - \\phi'(x)(y-x)), & y < x \\\\
+        \\alpha(\\phi(y) - \\phi(x) - \\phi'(x)(y-x)), & x \\leq y
+        \\end{cases}
+
+    where
+        - :math:`x` is the forecast
+        - :math:`y` is the observation
+        - :math:`\\alpha` is the expectile level
+        - :math:`\\phi` is a convex function of a single variable
+        - :math:`\\phi'` is the subderivative of :math:`\\phi`
+        - :math:`S(x,y)` is the score.
+
+    Note that if :math:`\\phi` is differentiable then :math:`\\phi'` is its derivative.
+
+    The score is negatively oriented with :math:`0 \\leq S(x,y) < \\infty`.
+
+    We strongly recommend that you work through the tutorial 
+    https://scores.readthedocs.io/en/stable/tutorials/Consistent_Scores.html to help you 
+    understand how to use the consistent scoring functions in `scores`.
+
+    Args:
+        fcst: array of forecast values.
+        obs: array of corresponding observation values.
+        alpha: expectile level. Must be strictly between 0 and 1.
+        phi: a convex function on the real numbers, accepting a single array like argument.
+        phi_prime: a subderivative of `phi`, accepting a single array like argument.
+        reduce_dims: Optionally specify which dimensions to reduce when
+            calculating the consistent expectile score. All other dimensions will be preserved. As a
+            special case, 'all' will allow all dimensions to be reduced. Only one
+            of `reduce_dims` and `preserve_dims` can be supplied. The default behaviour
+            if neither are supplied is to reduce all dims.
+        preserve_dims: Optionally specify which dimensions to preserve when calculating
+            the consistent quantile score. All other dimensions will be reduced. As a special case, 'all'
+            will allow all dimensions to be preserved. In this case, the result will be in
+            the same shape/dimensionality as the forecast, and the errors will be the consistent quantile
+            score at each point (i.e. single-value comparison against observed), and the
+            forecast and observed dimensions must match precisely. Only one of `reduce_dims`
+            and `preserve_dims` can be supplied. The default behaviour if neither are supplied
+            is to reduce all dims.
+        weights: Optionally provide an array for weighted averaging (e.g. by area, by latitude,
+            by population, custom)
+
+    Returns:
[…]
