+++ docs/api.md
@@ -23,6 +23,9 @@
 .. autofunction:: scores.continuous.correlation
 .. autofunction:: scores.continuous.multiplicative_bias
 .. autofunction:: scores.continuous.isotonic_fit
+.. autofunction:: scores.continuous.consistent_expectile_score
+.. autofunction:: scores.continuous.consistent_quantile_score
+.. autofunction:: scores.continuous.consistent_huber_score
 ```
 
 ## scores.probability
