### 2024-06-26 f067cf5
Update docstring rendering noting issues post-merge
