### 2024-06-26 1a36026
Update DOI to the journal reference for Diebold 1995
