+++ docs/paper.bib
@@ -109,7 +109,7 @@
 	number		 = {3},
 	pages		 = {253--263},
 	year		 = {1995},
-	doi          = {10.3386/t0169}
+	doi          = {10.1080/07350015.1995.10524599}
 }
 @Article{Brown:2021,
 	  author = "Barbara Brown and Tara Jensen and John Halley Gotway and Randy Bullock and Eric Gilleland and Tressa Fowler and Kathryn Newman and Dan Adriaansen and Lindsay Blank and Tatiana Burek and Michelle Harrold and Tracy Hertneky and Christina Kalb and Paul Kucera and Louisa Nance and John Opatz and Jonathan Vigh and Jamie Wolff",
