### 2024-06-26 6725fa2
Address a grammatical issue in the API docstring for true positive rate
