+++ src/scores/emerging/block_bootstrap/block_bootstrap.py
@@ -160,237 +160,49 @@ def block_bootstrap(
           the underlying chunking strategy of non-bootstrapped dims.
 
     """
-    # validate inputs
     if iterations <= 0:
         ValueError("`iterations` must be greater than 0.")
 
-    if len(block_sizes) != len(bootstrap_dims):
-        ValueError("`block_sizes` must be the same size as `bootstrap_dims`.")
-
     raise NotImplementedError("`block_bootstrap` is currently a stub.")
 
-    # reorder dimensions to align across all arrays
-    (arrs_reordered, all_arr_dims_ord) = reorder_all_arr_dims(
+    # note: this will re-organize the arrays
+    array_info_cln = ArrayInfoCollection.make_from_arrays(
         arrs,
         bootstrap_dims,
-        auto_order_unspecified_dims,
+        block_sizes,
+        fit_blocks_method,
+        auto_order_missing,
     )
 
-    axi_collection = make_axis_info_collection(
-        arrs,
-        bootstrap_dims=bootstrap_dims,
-        block_sizes=block_sizes,
-        fit_blocks_method=fit_blocks_method,
+    block_sampler = BlockSampler(
+        array_info_collection=array_info_cln,
+        cyclic=cyclic,
     )
 
-    # pre-generate samples for each axis
-    ax_block_indices = _sample_block_indices(list(axi_collection.iter()), cyclic)
-    ax_block_mapping = {d: ax_block_indices[i] for i, d in enumerate(axi_collection.dims_order)}
-    arrs_bootstrapped = []
+    # ---
+    # broadcast block sampling onto multiple iterations...
+    #
+    # TODO: for parallel operations, iter_arr should be chunked appropriately,
+    # unless the underlying data-arrays themselves are dask arrays, in which
+    # case parallelization happens through broadcast chunks.
+    iter_arr = xr.DataArray(range(iterations), dims="iterations")
 
-    for arr in arrs_reordered:
-        ax_block_indices_for_arr = [ax_block_mapping[d] for d in arr.dims]
-        pass
-        # TODO:
-        # - perform ufunc on `_construct_block_bootstrap_array` with core dims as bootstrap_dims
-        # - expand to `iterations`
-        # - append to result list.
+    def _block_sample_ufunc_wrapper(iter_arr_, *arrs_, block_sampler_):
+        return np.array([block_sampler_.sample_blocks_unchecked(list(arrs_)) for _ in iter_arr_])
+
+    # ---
+
+    # bootstrap dimensions should not be broadcast as they will be resampled as a block
+    # and will not abide by vectorization/chunking rules.
+    bootstrap_dims_per_arr = [arr.bootstrap_dims for arr in array_info_cln.array_info]
+
+    arrs_bootstrapped = xr.apply_ufunc(
+        iter_arr,
+        *tuple(array_info_cln.arrays_ordered),
+        core_input_dims=[[], *bootstrap_dims_per_arr],
+        output_core_dims=bootstrap_dims_per_arr,
+        kwargs={"block_sampler_": block_sampler},
+        dask="parallelized",  # TODO: get from keyword args
+    )
 
     return arrs_bootstrapped
-
-
-def _construct_block_bootstrap_array(
-    input_arr: npt.NDArray,
-    block_sizes: list[int],
-    cyclic: bool = True,
-    fit_blocks_method: FitBlocksMethod = FitBlocksMethod.PARTIAL,
-) -> npt.NDArray:
-    """
-    takes a numpy array and performs block resampling for 1 iteration,
-
-    note that the output array size is not guarenteed to be the same shape,
-    depending on FitBlocksMethod
-    """
-    assert len(block_sizes) == len(input_arr.shape)
-
-    # construct axis info based on block sizes and input array
[…]
