### 2024-06-28 nci/scores#574 @calebweinreb
[JOSS] Installation of jupyer kernel
Re JOSS review https://github.com/openjournals/joss-reviews/issues/6889

It would be helpful to include a step in the installation instructions for making the virtual environment available within jupyter, e.g.,

```
python -m ipykernel install --user --name=<env_name>
```
