### 2024-06-28 nci/scores#575 @calebweinreb
[JOSS] Instructions for downloading example data
Re JOSS review https://github.com/openjournals/joss-reviews/issues/6889

Currently the tutorial notebook `Additive_and_multiplicative_bias` includes the following note:
```
Note: In this tutorial we use the forecast and analysis grids that are downloaded or derived in First_Data_Fetching.ipynb. Please run through this tutorial first to fetch data.
```

This is very helpful and should ideally be added to other notebooks that use the same datasets, since users may not go through these notebooks in chronological order. For example the MSE, MAE, and RMSE notebooks all rely on `forecast_grid.nc` and `analysis_grid.nc` but don't explain where they come from.
