### 2024-06-28 nci/scores#576 @calebweinreb
[JOSS] General explanation of reduce/preserve
Re JOSS review openjournals/joss-reviews#6889

One of the most powerful features of `scores` is its flexible handling of multidimensional arrays, e.g., through the inclusion of "reduce" and "preserve" options for many (most?) of the supplied metrics. However, I think the correct usage of these options could be confusing to a new user, especially one who isn't already family with the concept of "reduce" in a more general programming context. The docstrings help with this a little but may still confuse some users since they also assume familiarity with the term "reduce".

To address this, I would advocate including a section on the documentation that describes "reduce/preserve" in general terms to a naive audience. This could take the form of a tutorial that explores how the options could be used in different multi-dim scenarios, or simply a text/diagrammatic explanation along the lines of

"Most of the metrics in scores are designed to compare groups of numbers. Sometimes it is useful to consider all the numbers in a multi-dim array as belonging so one group. Other times it is useful to group along a subset of dimensions and then apply the metric repeatedly across the non-grouped dimensions..." etc...
