### 2024-06-28 nci/scores#577 @calebweinreb
[JOSS] Minimal pandas support
Re JOSS review openjournals/joss-reviews#6889

There currently appears to be a mismatch between the level of pandas support implied by the JOSS manuscript and that which is actually available in the scores package. (Apologies in advance if I have overlooked any pandas options already in the package).

- Currently, support for pandas is advertised at several points in the JOSS manuscript, including the title, the compatibility section, and as an advantage over existing packages (e.g., xskillscore).
- However as far as I can tell, the pandas support remains truly minimal at this point. Of the 50 metrics supplied by scores, only 3 are supported in pandas (MAE, MSE, RMSE), and all of them are routine calculations that could be obtained through many other packages. Furthermore, with the exception of "is_angular", many of the useful option in scores (weighting, preserve/reduce) are not afforded through the pandas API.
- Though the JOSS manuscript does describe the pandas support as "expanding", this feels somewhat misleading given how little is currently implemented.

I think this could be addressed in two ways.

- One option would be to deemphasize pandas support in the JOSS manuscript, e.g., by removing it from the title and in the comparison to alternative packages. Of course, it's awesome that pandas support is on the way -- and this should definitely mentioned -- but in the mean time, one would want to avoid pandas users reading the JOSS paper and falsely concluding that they can use the more non-trivial metrics offered by scores.
- Another option would be to temporarily shoehorn pandas support into the rest of the API by wrapping the existing metrics with functions that convert (pandas -> xarray) on the input side and (xarray -> pandas) on the output side (or something like that).
