### 2024-06-28 nci/scores#578 @calebweinreb
[JOSS] Implementation of weights is occasionally unclear
Re JOSS review openjournals/joss-reviews#6889

A big strength of `scores` is the ability to incorporate weights into many of the metrics. But how exactly these weights are incorporated is often unclear. Though the application of weights is obvious in come cases (e.g. RMSE), I don't have any intuition for how weights be used `roc_curve_data`, for example. In these ambiguous cases the docstring should be updated to provide a clearer explanation.
