### 2024-06-28 #6889 @calebweinreb
@tennlee @crvernon

I have gone through the paper/docs and installed/tested the package. Overall I am super impressed with the documentation and the completeness and polish of the package. I raised a few issues related to the manuscript and documentation. The only blocking issue is https://github.com/nci/scores/issues/577, which must be addressed or rebutted before I am willing to approve "summary" and "statement of need" in the checklist. https://github.com/nci/scores/issues/575 and https://github.com/nci/scores/issues/574 seem like easy wins to me. I think https://github.com/nci/scores/issues/576 and https://github.com/nci/scores/issues/578 could improve clarity/accessibility of the tool, but as an outsider to the field, I am not familiar with the target audience and therefore defer to the developers on whether these additions would be broadly helpful.
