### 2024-06-29 344f6fa
Potential updates to JOSS Paper
