### 2024-06-30 44fd5a1
Update bib.md entry
