### 2024-06-30 nci/scores#578 @tennlee
That makes sense. It will take some time, but I will work through each score/metric and check if there is an example of the weights presented in each notebook as well as checking the docstrings to provide an indication.
