### 2024-06-30 nci/scores#576 @tennlee
Thanks for this feedback. Most of the current notebooks are very much 'per score' rather than 'per concept'. It will take some time, but I'll happily create a tutorial which specifically addresses the purpose and use of reduce/preserve. The MAE and RMSE tutorials do demonstrate this briefly, but I agree some more explanation would be useful for many.
