### 2024-06-30 nci/scores#575 @tennlee
Thanks for that feedback. I have addressed it in PR #579 . Feel free to re-open this issue if you have any further feedback.
