### 2024-06-30 a67b465
Further update to paper.bib entry
