+++ docs/paper.bib
@@ -234,7 +234,7 @@
       eprint={2404.18429},
       archivePrefix={arXiv},
       primaryClass={physics.ao-ph},
-      publisher={arXiv preprint arXiv:2404.18429},
+      journal={arXiv preprint arXiv:2404.18429},
 	  doi={10.48550/arXiv.2404.18429}
 }
 @article{Ferro_2013, 
