### 2024-06-30 bbe0c54
Another tweak to paper.bib
