### 2024-06-30 11002ba
Add version pins based on somewhat recent rather than latest versions to allow more flexibility for users
