+++ pyproject.toml
@@ -20,11 +20,11 @@ classifiers = [
     "Operating System :: OS Independent",
 ]
 dependencies = [
-    "xarray",
-    "pandas",
-    "scipy",
-    "bottleneck",
-    "scikit-learn",
+    "xarray ~= 2024",
+    "pandas ~= 2.2",
+    "scipy" ~= 1.1",
+    "bottleneck ~= 1.3",
+    "scikit-learn ~= 1.4",
 ]
 
 [project.optional-dependencies]
