+++ pyproject.toml
@@ -22,7 +22,7 @@ classifiers = [
 dependencies = [
     "xarray ~= 2024",
     "pandas ~= 2.2",
-    "scipy" ~= 1.1",
+    "scipy ~= 1.1",
     "bottleneck ~= 1.3",
     "scikit-learn ~= 1.4",
 ]
