### 2024-06-30 b105587
Test alternative version number for xarray dependency The previous code worked on local testing but failed in the pipeline
