### 2024-06-30 d0011ce
Further paper updates
