### 2024-06-30 d5d007c
Change pysteps capitalisation
