### 2024-06-30 7ac8bd2
JOSS paper update (#581)
* Responds to reviewer feedback
