+++ docs/paper.bib
@@ -227,14 +227,14 @@
   publisher={CSIRO Publishing},
   doi={10.1071/es21010}
 }
-@misc{loveday2024jive,
+@article{loveday2024jive,
       title={{The Jive verification system and its transformative impact on weather forecasting operations}}, 
       author={Nicholas Loveday and Deryn Griffiths and Tennessee Leeuwenburg and Robert Taggart and Thomas C. Pagano and George Cheng and Kevin Plastow and Elizabeth Ebert and Cassandra Templeton and Maree Carroll and Mohammadreza Khanarmuei and Isha Nagpal},
       year={2024},
       eprint={2404.18429},
       archivePrefix={arXiv},
       primaryClass={physics.ao-ph},
-      publisher={arXiv},
+      journal={{arXiv {p}reprint arXiv:2404.18429}},
 	  doi={10.48550/arXiv.2404.18429}
 }
 @article{Ferro_2013, 
