### 2024-06-30 nci/scores#577 @tennlee
@calebweinreb we have merged a change to the paper into 528-joss-paper-submission which we believe addresses your feedback. Please let us know whether this seems suitable.
