### 2024-06-30 #6889 @tennlee
@calebweinreb here is a revised version of the paper which we believe addresses your feedback on https://github.com/nci/scores/issues/577. Let us know if this seems suitable.
