### 2024-07-01 dbcfd89
Version pinning for main (#580)
* Add version pins based on somewhat recent rather than very latest versions to allow more flexibility for users
* Manually ran unit tests on an environment with the oldest versions installed in it
* Update required Python version 3.9
