+++ pyproject.toml
@@ -13,18 +13,18 @@ Scores is a package containing mathematical functions \
 for the verification, evaluation and optimisation of forecasts, predictions or models.
 """
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.9"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: Apache Software License",
     "Operating System :: OS Independent",
 ]
 dependencies = [
-    "xarray",
-    "pandas",
-    "scipy",
-    "bottleneck",
-    "scikit-learn",
+    "xarray ~= 2024.1",
+    "pandas ~= 2.2",
+    "scipy ~= 1.1",
+    "bottleneck ~= 1.3",
+    "scikit-learn ~= 1.4",
 ]
 
 [project.optional-dependencies]
