### 2024-07-01 nci/scores#577 @calebweinreb
The new draft looks great and totally addresses my concern.
