### 2024-07-02 a2e6dc6
minor updates to pearsonr
