### 2024-07-02 5b62d2d
rename correlation as pearsonr (#583)
* rename correlation as pearsonr, place within scores.continuous.correlation

---------

Signed-off-by: Nicholas Loveday <48701367+nicholasloveday@users.noreply.github.com>
