### 2024-07-02 5b47bdc
Add documentation for the registration of `scores` as an additional Jupyter kernel
