### 2024-07-02 8557cef
Add installation documentation for Jupyter kernels
