### 2024-07-02 31afd72
Installation documentation for Jupyter kernel options (#586)
* Add documentation for the registration of Jupyter kernels
