### 2024-07-02 nci/scores#574 @tennlee
@calebweinreb I have merged PR #586  which addresses this issue. Please feel free to comment or re-open the issue if you have any further feedback.
