### 2024-07-02 nci/scores#576 @tennlee
I would like to clarify my earlier comments. I think such a notebook could be of particular value for those outside or new to the field, which is what I meant when I said "more explanation would be useful for many". As such, down the road, I do intend to add a new tutorial specifically about these concepts.  In the meantime, there are some examples in the existing tutorials.

I think the main audience for `scores` will already be familiar with the general concept of dimensionality reduction, and be fairly comfortable using the preserve/reduce functionality.
