### 2024-07-02 #6889 @tennlee
@calebweinreb we have now resolved [nci/scores#577](https://github.com/nci/scores/issues/577), which you said earlier was the only blocking issue. We have also resolved [nci/scores#575](https://github.com/nci/scores/issues/575) and [nci/scores#574](https://github.com/nci/scores/issues/574) which you raised.

As noted in issue [nci/scores#578](https://github.com/nci/scores/issues/578), this will need more time to resolve. My understanding is that you do not believe it to be a blocking issue.

Regarding [nci/scores#576](https://github.com/nci/scores/issues/576), I have posted a clarification on the issue thread, to the effect that while it would be helpful for some new users or those outside the field, I don’t think it is an issue for the main audience of `scores`. As such, I do intend to address it, but I believe it can be done subsequent to publication. This will allow more time for the development and review of the tutorial.

Can you please let me know whether you are happy for publication to proceed as-is, or whether [nci/scores#576](https://github.com/nci/scores/issues/576) should be addressed first.
