### 2024-07-03 73f11b7
Improve clarity in overview and Jupyter Notebook section
