### 2024-07-03 c4448cb
Installation Options Update to improve clarity in overview and Jupyter Notebook sections (#587)
* Improve clarity in overview and Jupyter Notebook section
* Improve clarity in third dot point in overview
