### 2024-07-03 767c37c
Update installation.md
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
