### 2024-07-03 49907c8
Update installation.md (#588)
* Update installation.md

---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
