### 2024-07-03 #6889 @tennlee
@calebweinreb We have now added a tutorial regarding dimensions handling, to address your issue 576 (reduce/preserve handling).

I think this now addresses everything required.
