### 2024-07-07 aa130e6
Update maintainers notes with release note template information
