### 2024-07-07 9f1e29b
Update one sentence for clarity
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
