### 2024-07-07 18e8697
Update header in Release Notes template
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
