### 2024-07-07 3594d52
Define AUC on first use. Re-order docstring so that the html formats correctly.
