### 2024-07-07 e447d21
Better align the paragraph text, also improves rendering
