### 2024-07-07 3a5a368
Add release note template information to maintainers notes (#593)
* Update maintainers notes with release note template information

* Update header in Release Notes template

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>

---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
