### 2024-07-08 9bddd60
Remove one word
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
