+++ docs/release_notes.md
@@ -2,7 +2,7 @@
 
 ## Version 0.9.3 (Upcoming Release)
 
-For the full details of all changes in this release, see the GitHub commit history. Below are the changes we think users may wish to be aware of.
+For the full details of all changes in this release, see the [GitHub commit history](https://github.com/nci/scores/compare/0.9.2...develop). Below are the changes we think users may wish to be aware of.
 
 ### Breaking Changes
 
