### 2024-07-08 ca20330
Adds Release Notes page to documentation (#592)
* Adds Release Notes page to documentation

---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
