+++ docs/release_notes.md
@@ -1,0 +1,19 @@
+# Release Notes (What's New)
+
+## Version 0.9.3 (Upcoming Release)
+
+For the full details of all changes in this release, see the [GitHub commit history](https://github.com/nci/scores/compare/0.9.2...develop). Below are the changes we think users may wish to be aware of.
+
+### Breaking Changes
+
+- Renamed and relocated function `scores.continuous.correlation` to `scores.continuous.correlation.pearsonr`. See [PR #583](https://github.com/nci/scores/pull/583) by [@nicholasloveday](https://github.com/nicholasloveday). 
+
+### Documentation
+
+- Added "Dimension Handling" tutorial, which describes reducing and preserving dimensions. See [PR #589](https://github.com/nci/scores/pull/589) by [@nicholasloveday](https://github.com/nicholasloveday).
+- Updated "Detailed Installation Guide" with information on installing kernels in a Jupyter environment. See [PR #586](https://github.com/nci/scores/pull/586) by [@tennlee](https://github.com/tennlee) and [PR #587](https://github.com/nci/scores/pull/587) by [@Steph-Chong](https://github.com/Steph-Chong).
+
+### Internal Changes
+
+- Introduced pinned versions for dependencies on main. See [PR #580](https://github.com/nci/scores/pull/580)  by [@tennlee](https://github.com/tennlee).
+
