### 2024-07-08 008f3f5
Minor update to template for release notes
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
