### 2024-07-08 9709615
Hopefully add proper finalisation of coveralls testing in GitHub workflow
