+++ .github/workflows/python-app.yml
@@ -41,4 +41,10 @@ jobs:
       with:
         flag-name: run-${{ join(matrix.*, '-') }}
         parallel: true
+  finish:
+    needs: build
+    steps:
+      - name: Indicate Coveralls Parallel Finished
+        uses: coverallsapp/github-action@v2.3.0
+        with: parallel-finished: true
   
