+++ .github/workflows/python-app.yml
@@ -46,5 +46,6 @@ jobs:
     steps:
     - name: Indicate Coveralls Parallel Finished
       uses: coverallsapp/github-action@v2.3.0
-      with: parallel-finished: true
+      with: 
+        parallel-finished: true
   
