+++ .github/workflows/python-app.yml
@@ -44,7 +44,6 @@ jobs:
   finish:
     needs: build
     runs-on: ubuntu-latest
-    if: $
     steps:
     - name: Indicate Coveralls Parallel Finished
       uses: coverallsapp/github-action@v2.3.0
