### 2024-07-08 57e02f7
Minor tweaks to release note formatting section of Maintainer's Notes
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
