+++ docs/maintainer.md
@@ -44,7 +44,7 @@ Information relevant for package maintenance
 ```
 # Release Notes (What's New)
 
-## Version X.Y.Z (Month Day, Year) e.g. "0.9.3 (July 7, 2024)"
+## Version X.Y.Z (Month Day, Year) e.g. "Version 0.9.3 (July 7, 2024)"
 
 For the full details of all changes in this release, see the [GitHub commit history](https://github.com/nci/scores/compare/(X-1).(Y-1).(Z-1)...X.Y.Z). Below are the changes we think users may wish to be aware of."
 
@@ -55,7 +55,7 @@ For the full details of all changes in this release, see the [GitHub commit hist
 ### Documentation
 ### Internal Changes
 
-For each entry: "Brief description. See [PR #XYX](add link) by [@username](add link)."
+For each entry: "Brief description. See [PR #ABC](add link) by [@username](add link)."
 ```
 
 ## This section covers how to build the documentation locally 
