### 2024-07-08 aaa2557
Update CITATION.cff
Update citation to match updated paper

Signed-off-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
