### 2024-07-08 #6889 @crvernon
👋  @tennlee - In your paper's references, please ensure that capitalization has been maintained where it is necessary.   You can do this using curly brackets around the formatting you want to maintain in your bib file.  For example, in LINE 219 in the paper, "...with generalized huber loss..." the "h" in "huber" should be capitalized.  Please check other refs for this as well.

So here is what we have left to do:

- [ ]  Please respond with the DOI of the archived version here
