### 2024-07-09 281fe0d
Fine-grained adjustments to citations
