+++ docs/paper.bib
@@ -155,7 +155,7 @@
 	doi          = {10.1002/qj.4206}
 }
 @Article{Taggart:2022c,
-	title		 = {Point forecasting and forecast evaluation with generalized Huber loss},
+	title		 = {Point forecasting and forecast evaluation with generalized {H}uber loss},
 	author		 = {Taggart, Robert},
 	journal		 = {Electronic Journal of Statistics},
 	volume		 = {16},
@@ -237,19 +237,19 @@
       journal={{arXiv {p}reprint arXiv:2404.18429}},
 	  doi={10.48550/arXiv.2404.18429}
 }
-@article{Ferro_2013, 
-	title={Fair scores for ensemble forecasts: Fair Scores for Ensemble Forecasts}, 
-	volume={140}, 
-	ISSN={0035-9009}, 
-	url={http://dx.doi.org/10.1002/qj.2270}, 
-	DOI={10.1002/qj.2270}, 
-	number={683}, 
-	journal={Quarterly Journal of the Royal Meteorological Society}, 
-	publisher={Wiley}, 
-	author={Ferro, C. A. T.}, 
-	year={2013}, 
-	month=dec, 
-	pages={1917–1923}
+@article{Ferro_2014, 
+	author = {Ferro, C. A. T.},
+	title = {Fair scores for ensemble forecasts},
+	journal = {Quarterly Journal of the Royal Meteorological Society},
+	volume = {140},
+	number = {683},
+	pages = {1917-1923},
+	keywords = {Brier score, continuous ranked probability score, scoring rules, forecast verification},
+	doi = {10.1002/qj.2270},
+	url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/qj.2270},
+	eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/qj.2270},
+	abstract = {Abstract The notion of fair scores for ensemble forecasts was introduced recently to reward ensembles with members that behave as though they and the verifying observation are sampled from the same distribution. In the case of forecasting binary outcomes, a characterization is given of a general class of fair scores for ensembles that are interpreted as random samples. This is also used to construct classes of fair scores for ensembles that forecast multicategory and continuous outcomes. The usual Brier, ranked probability and continuous ranked probability scores for ensemble forecasts are shown to be unfair, while adjusted versions of these scores are shown to be fair. A definition of fairness is also proposed for ensembles with members that are interpreted as being dependent and it is shown that fair scores exist only for some forms of dependence.},
+	year = {2014}
 }
 @article{Gneiting_2007, 
 	title={Strictly Proper Scoring Rules, Prediction, and Estimation}, 
@@ -292,7 +292,7 @@
 }
 @software{pandas:2024,
   author       = {{The pandas development team}},
-  title        = {pandas-dev/pandas: Pandas},
+  title        = {{pandas-dev/pandas: Pandas}},
   month        = apr,
   year         = 2024,
   publisher    = {Zenodo},
@@ -318,7 +318,7 @@
 @software{Morley:2020,
   author       = {Steve Morley and
                   Angeline Burrell},
-  title        = {drsteve/PyForecastTools: Version 1.1.1},
+  title        = {{drsteve/PyForecastTools: Version 1.1.1}},
   month        = apr,
   year         = 2020,
   publisher    = {Zenodo},
@@ -342,41 +342,44 @@
   year = {2020},
   month = {feb}
 }
-@software{zarr:2020,
+@software{zarr:2024,
   author       = {Alistair Miles and
-                  John Kirkham and
-                  Martin Durant and
-                  James Bourbeau and
-                  Tarik Onalan and
-                  Joe Hamman and
-                  Zain Patel and
-                  shikharsg and
[…]
