### 2024-07-09 b68de70
Update citation from misc to software and add version
