+++ docs/paper.bib
@@ -24,13 +24,14 @@
 	eprint       = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/met.1732},
 	abstract     = {The degree to which a forecast changes from one issue time to the next is an interesting aspect of a forecast system. Weather forecasters report that they are reluctant to change a forecast if they judge there is a risk of it being changed back again. They believe such instability detracts from the message being delivered and are reluctant to use automated guidance which they perceive as having lack of stability. A Flip-Flop Index was developed to quantify this characteristic of revisions of fixed-event forecasts. The index retains physically meaningful units, has a simple definition and does not penalize a sequence of forecasts that show a trend, which is important when assessing forecasts where a trend can be interpreted as a forecast becoming more confident with a shorter lead time. The Flip-Flop Index was used to compare the stability of sequences of automated guidance with the official Australian Bureau of Meteorology forecasts, which are prepared manually. The results show that the forecasts for chance of rain from the automated guidance are often more stable than the official, manual forecasts. However, the official forecasts for maximum temperature are more stable than those based on automated guidance. The Flip-Flop Index is independent of observations and does not measure skill, but it can play a complementary role in characterizing and evaluating a forecasting system.}
 }
-@misc{xskillscore,
+@software{xskillscore,
 	title        = {{xarray-contrib/xskillscore: Metrics for verifying forecasts}},
 	author       = {Ray Bell and Aaron Spring and Riley Brady and Andrew Huang and Dougie Squire and Zachary Blackwood and Maximillian Cosmo Sitter and Taher Chegini.},
 	month        = aug,
 	year         = 2021,
 	publisher    = {Zenodo},
-	doi          = {10.5281/zenodo.5173153},	
+	doi          = {10.5281/zenodo.5173153},
+	version      = {v0.0.23},
 	url          = {https://github.com/xarray-contrib/xskillscore}
 }
 @article{Ehm:2016,
@@ -359,7 +360,7 @@
                   Mads R. B. Kristensen and
                   Norman Rzepka and
                   Saransh Chopra and
-                  AWA BRANDON AWA and
+                  Awa Brandon Awa and
                   Sanket Verma and
                   Matthew Rocklin and
                   Max Jones and
