### 2024-07-09 1b8f88b
Fine-grained adjustments to citations (#597)
* Fine-grained adjustments to citations
* Update citation from misc to software and add version
