+++ docs/paper.bib
@@ -24,13 +24,14 @@
 	eprint       = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/met.1732},
 	abstract     = {The degree to which a forecast changes from one issue time to the next is an interesting aspect of a forecast system. Weather forecasters report that they are reluctant to change a forecast if they judge there is a risk of it being changed back again. They believe such instability detracts from the message being delivered and are reluctant to use automated guidance which they perceive as having lack of stability. A Flip-Flop Index was developed to quantify this characteristic of revisions of fixed-event forecasts. The index retains physically meaningful units, has a simple definition and does not penalize a sequence of forecasts that show a trend, which is important when assessing forecasts where a trend can be interpreted as a forecast becoming more confident with a shorter lead time. The Flip-Flop Index was used to compare the stability of sequences of automated guidance with the official Australian Bureau of Meteorology forecasts, which are prepared manually. The results show that the forecasts for chance of rain from the automated guidance are often more stable than the official, manual forecasts. However, the official forecasts for maximum temperature are more stable than those based on automated guidance. The Flip-Flop Index is independent of observations and does not measure skill, but it can play a complementary role in characterizing and evaluating a forecasting system.}
 }
-@misc{xskillscore,
+@software{xskillscore,
 	title        = {{xarray-contrib/xskillscore: Metrics for verifying forecasts}},
 	author       = {Ray Bell and Aaron Spring and Riley Brady and Andrew Huang and Dougie Squire and Zachary Blackwood and Maximillian Cosmo Sitter and Taher Chegini.},
 	month        = aug,
 	year         = 2021,
 	publisher    = {Zenodo},
-	doi          = {10.5281/zenodo.5173153},	
+	doi          = {10.5281/zenodo.5173153},
+	version      = {v0.0.23},
 	url          = {https://github.com/xarray-contrib/xskillscore}
 }
 @article{Ehm:2016,
@@ -155,7 +156,7 @@
 	doi          = {10.1002/qj.4206}
 }
 @Article{Taggart:2022c,
-	title		 = {Point forecasting and forecast evaluation with generalized Huber loss},
+	title		 = {Point forecasting and forecast evaluation with generalized {H}uber loss},
 	author		 = {Taggart, Robert},
 	journal		 = {Electronic Journal of Statistics},
 	volume		 = {16},
@@ -237,19 +238,19 @@
       journal={{arXiv {p}reprint arXiv:2404.18429}},
 	  doi={10.48550/arXiv.2404.18429}
 }
-@article{Ferro_2013, 
-	title={Fair scores for ensemble forecasts: Fair Scores for Ensemble Forecasts}, 
-	volume={140}, 
-	ISSN={0035-9009}, 
-	url={http://dx.doi.org/10.1002/qj.2270}, 
-	DOI={10.1002/qj.2270}, 
-	number={683}, 
-	journal={Quarterly Journal of the Royal Meteorological Society}, 
-	publisher={Wiley}, 
-	author={Ferro, C. A. T.}, 
-	year={2013}, 
-	month=dec, 
-	pages={1917–1923}
+@article{Ferro_2014, 
+	author = {Ferro, C. A. T.},
+	title = {Fair scores for ensemble forecasts},
+	journal = {Quarterly Journal of the Royal Meteorological Society},
+	volume = {140},
+	number = {683},
+	pages = {1917-1923},
+	keywords = {Brier score, continuous ranked probability score, scoring rules, forecast verification},
+	doi = {10.1002/qj.2270},
+	url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/qj.2270},
+	eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/qj.2270},
[…]
