### 2024-07-09 #6889 @Steph-Chong
@crvernon we have capitalized the "h" in "Huber" in our references section. We have also done another check of the references section. Please let us know if there is anything else that needs our attention.

We still need to make the GitHub release and archival deposit. Once that is done, we will post here with the version number and archive DOI.
