### 2024-07-09 60d175a
Add date and updated URL to release notes (#598)
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
