### 2024-07-09 1e03a22
Set version numbers for next release
