### 2024-07-09 #6889 @tennlee
We have proceeded to release version 0.9.3. I have also updated the branch containing paper.md to match. Release 0.9.3 has been lodged with Zenodo. The DOI link is https://doi.org/10.5281/zenodo.12697242 . This link is not yet resolving but will presumably update shortly. A direct link to the zenodo record is https://zenodo.org/records/12697242 . I have gone through the checklist you posted, but I was unsure if I should tick that list off myself or whether you would like to double-check.
