### 2024-07-09 0b0e95d
Draft release notes for version 1.0.0
