### 2024-07-09 2624a6e
Update release note wording
