### 2024-07-09 #6889 @savente93
You're welcome! I'm very glad to hear you found it a constructive process. Congrats, and well done on the publication 🎉
