### 2024-07-10 42187ed
Update citation to refer to the JOSS paper
