+++ CITATION.cff
@@ -32,6 +32,8 @@ authors:
   given-names: John
 message: If you use this software, please cite our article in the
   Journal of Open Source Software.
+title: "scores: A Python package for verifying and evaluating models and
+  predictions with xarray"  
 preferred-citation:
   authors:
   - family-names: Leeuwenburg
@@ -77,5 +79,4 @@ preferred-citation:
   type: article
   url: "https://joss.theoj.org/papers/10.21105/joss.06889"
   volume: 9
-title: "scores: A Python package for verifying and evaluating models and
-  predictions with xarray"
+
