### 2024-07-10 7093294
Update CITATION.cff to refer to the JOSS paper (#602)
* Update citation to refer to the JOSS paper
