+++ CITATION.cff
@@ -1,73 +1,82 @@
-cff-version: 1.2.0
-message: "If you use this software, please cite it as below."
+cff-version: "1.2.0"
 authors:
-- given-names: Tennessee 
-  family-names: Leeuwenburg
+- family-names: Leeuwenburg
+  given-names: Tennessee
   orcid: "https://orcid.org/0009-0008-2024-1967"
-- given-names: Nicholas 
-  family-names: Loveday
+- family-names: Loveday
+  given-names: Nicholas
   orcid: "https://orcid.org/0009-0000-5796-7069"
-- given-names: Elizabeth E. 
-  family-names: Ebert
-- given-names: Harrison 
-  family-names: Cook
+- family-names: Ebert
+  given-names: Elizabeth E.
+- family-names: Cook
+  given-names: Harrison
   orcid: "https://orcid.org/0009-0009-3207-4876"
-- given-names: Mohammadreza 
-  family-names: Khanarmuei
+- family-names: Khanarmuei
+  given-names: Mohammadreza
   orcid: "https://orcid.org/0000-0002-5017-9622"
-- given-names: Robert J. 
-  family-names: Taggart
+- family-names: Taggart
+  given-names: Robert J.
   orcid: "https://orcid.org/0000-0002-0067-5687"
-- given-names: Nikeeth 
-  family-names: Ramanathan
+- family-names: Ramanathan
+  given-names: Nikeeth
   orcid: "https://orcid.org/0009-0002-7406-7438"
-- given-names: Maree 
-  family-names: Carroll
+- family-names: Carroll
+  given-names: Maree
   orcid: "https://orcid.org/0009-0008-6830-8251"
-- given-names: Stephanie 
-  family-names: Chong
+- family-names: Chong
+  given-names: Stephanie
   orcid: "https://orcid.org/0009-0007-0796-4127"
-- given-names: Aidan 
-  family-names: Griffiths
-- given-names: John 
-  family-names: Sharples  
-
-title: "scores: A Python package for verifying and evaluating models and predictions with xarray"  
-
+- family-names: Griffiths
+  given-names: Aidan
+- family-names: Sharples
+  given-names: John
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+title: "scores: A Python package for verifying and evaluating models and
+  predictions with xarray"  
 preferred-citation:
-  type: article
-  doi: "10.48550/arXiv.2406.07817"
-  journal: "arXiv"
-  title: "scores: A Python package for verifying and evaluating models and predictions with xarray"
   authors:
-  - given-names: Tennessee 
-    family-names: Leeuwenburg
+  - family-names: Leeuwenburg
+    given-names: Tennessee
     orcid: "https://orcid.org/0009-0008-2024-1967"
-  - given-names: Nicholas 
-    family-names: Loveday
+  - family-names: Loveday
+    given-names: Nicholas
     orcid: "https://orcid.org/0009-0000-5796-7069"
-  - given-names: Elizabeth E. 
-    family-names: Ebert
-  - given-names: Harrison 
-    family-names: Cook
+  - family-names: Ebert
+    given-names: Elizabeth E.
+  - family-names: Cook
+    given-names: Harrison
     orcid: "https://orcid.org/0009-0009-3207-4876"
-  - given-names: Mohammadreza 
-    family-names: Khanarmuei
+  - family-names: Khanarmuei
+    given-names: Mohammadreza
     orcid: "https://orcid.org/0000-0002-5017-9622"
-  - given-names: Robert J. 
-    family-names: Taggart
+  - family-names: Taggart
+    given-names: Robert J.
     orcid: "https://orcid.org/0000-0002-0067-5687"
[…]
