### 2024-07-10 f220e04
Update the readme to include citation information for the new paper (#603)
