### 2024-07-10 ca9422c
Add the "badge" for the JOSS paper (#604)
