### 2024-07-10 f17deca
Simplify the release notes
