### 2024-07-10 537fa67
Update version information for version 1.0.0
