### 2024-07-10 85a1004
Release notes for version 1.0.0 (#600)
* Draft release notes for version 1.0.0

---------

Signed-off-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
Co-authored-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
