### 2024-07-10 1dfccdf
Add paper link to README
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
