+++ README.md
@@ -71,7 +71,7 @@ All metrics, statistical techniques and data processing tools in `scores` work w
 
 ## Acknowledging This Work
 
-If you use `scores` for a published work, we would appreciate you citing our paper:
+If you use `scores` for a published work, we would appreciate you citing our [paper](https://doi.org/10.21105/joss.06889):
 
 Leeuwenburg, T., Loveday, N., Ebert, E. E., Cook, H., Khanarmuei, M., Taggart, R. J., Ramanathan, N., Carroll, M., Chong, S., Griffiths, A., & Sharples, J. (2024). scores: A Python package for verifying and evaluating models and predictions with xarray. *Journal of Open Source Software, 9*(99), 6889. [https://doi.org/10.21105/joss.06889](https://doi.org/10.21105/joss.06889)
 
