### 2024-07-10 43efac3
Add hyperlink to DOI in release notes
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
