### 2024-07-10 #6889 @tennlee
@crvernon I have noticed a possible error in the Crossref metadata. I am looking at https://api.crossref.org/works/10.21105/joss.06889 . The title has an errant newline instead of a space.

There seem to be newlines in the metadata in other places too, so perhaps it's just how the metadata is - perhaps Crossref have a line width requirement and perhaps it's just their formatting and not an error after all.

The "cite" tool in Crossref is generating citations that look correct - no sign of a newline in there.

What caused me to go looking was when I added the record to my ORCID, there was no space between "evaluating" and "models". This is the same location as the newline appears in the JSON from Crossref.

I'm not sure whether this is significant or not but I wanted to mention it.
