### 2024-07-10 #6889 @danielskatz
The metadata we submitted to Crossref is visible in https://github.com/openjournals/joss-papers/pull/5592/files, and it indeed has some strange newlines in it. I'm unsure about what happened here, and am pinging
