+++ src/scores/continuous/__init__.py
@@ -23,6 +23,12 @@ from scores.continuous.standard_impl import (
     rmse,
 )
 from scores.processing.isoreg_impl import isotonic_fit
+from scores.continuous.threshold_weighted_scores import (
+    threshold_weighted_score,
+    threshold_weighted_quantile_score,
+    threshold_weighted_absolute_error,
+    threshold_weighted_squared_error,
+)
 
 __all__ = [
     "flip_flop_index",
@@ -41,4 +47,8 @@ __all__ = [
     "consistent_expectile_score",
     "consistent_huber_score",
     "consistent_quantile_score",
+    "threshold_weighted_score",
+    "threshold_weighted_quantile_score",
+    "threshold_weighted_absolute_error",
+    "threshold_weighted_squared_error",
 ]
