+++ src/scores/continuous/threshold_weighted_scores.py
@@ -1,0 +1,647 @@
+"""
+Implementation of some basic threshold weighted scoring functions.
+See Taggart (2022) https://doi.org/10.1002/qj.4206
+"""
+
+import functools
+from typing import Callable, Literal, Optional, Sequence, Tuple, Union
+
+import numpy as np
+import xarray as xr
+
+from scores.typing import FlexibleDimensionTypes
+
+from scores.continuous.consistent_impl import (
+    check_alpha,
+    check_huber_param,
+    consistent_expectile_score,
+    consistent_huber_score,
+    consistent_quantile_score,
+)
+
+HUBER_FUNCS = ["huber_loss", "scaled_huber_loss"]
+ALPHA_FUNCS = ["quantile_score", "expectile_score"]
+OTHER_FUNCS = ["squared_error", "absolute_error"]
+SCORING_FUNCS = OTHER_FUNCS + ALPHA_FUNCS + HUBER_FUNCS
+AuxFuncType = Callable[[xr.DataArray], xr.DataArray]
+EndpointType = Union[int, float, xr.DataArray]
+
+
+def threshold_weighted_score(
+    fcst: xr.DataArray,
+    obs: xr.DataArray,
+    scoring_func: Literal[
+        "squared_error",
+        "absolute_error",
+        "quantile_score",
+        "expectile_score",
+        "huber_loss",
+        "scaled_huber_loss",
+    ],
+    interval_where_one: Tuple[EndpointType, EndpointType],
+    *,
+    interval_where_positive: Optional[Tuple[EndpointType, EndpointType]] = None,
+    alpha: Optional[float] = None,
+    huber_param: Optional[float] = None,
+    reduce_dims: Optional[FlexibleDimensionTypes] = None,
+    preserve_dims: Optional[FlexibleDimensionTypes] = None,
+    weights: Optional[xr.DataArray] = None,
+) -> xr.DataArray:
+    """
+    Returns the scores calculated using the specified threshold weighted scoring function,
+    averaged over the specified dimensions in `dims`.
+
+    Taggart (2022) showed that many commonly used scoring functions, such as the squared
+    error and absolute error scoring functions, can be decomposed as a sum of weighted
+    scoring functions, where the weights sum to 1, and where the original scoring function and
+    each member of its decomposition are consistent for the same statistical functional.
+    Each member of the decomposition measures predictive performance with emphasis on the
+    region specified by the corresponding weight.
+
+    The function `threshold_weighted_score` supports threshold weighting for the following
+    scoring functions :math:`S(x, y)`, where :math:`x` is the forecast and :math:`y` the observation:
+
+        - squared error (see `mse`): :math:`S(x, y) = (x - y)^2`
+        - absolute error: :math:`S(x, y) = |x - y|`
+        - quantile score (see `quantile_score`): :math:`S(x, y) = (1 - \\alpha)(x - y)`
+          if :math:`x \\geq y` and :math:`S(x, y) = \\alpha(y - x)` if :math:`x < y`
+        - expectile score: :math:`S(x, y) = (1 - \\alpha)(x - y)^2`
+          if :math:`x \\geq y` and :math:`S(x, y) = \\alpha(y - x)^2` if :math:`x < y`
+        - Huber loss: :math:`S(X, y) = \\frac\\{1\\}\\{2\\}(x-y)^2` if :math:`|x - y| \\leq a` and
+          :math:`S(X, y) = a|x-y| - a^2/2` if :math:`|x - y| > a`
+        - Scaled Huber loss: Huber loss scaled by a factor of :math:`1/a`
+
+    Two types of threshold weighting are supported: rectangular and trapezoidal.
+
+        - To specify a rectangular weight, set `interval_where_positive=None` and set
[…]
