### 2024-07-12 8cc8f66
added initial tests
