+++ tests/continuous/test_threshold_weighted.py
@@ -1,0 +1,656 @@
+"""
+Tests for scores.continuous.threshold_weighted_impl
+"""
+
+import numpy as np
+import pytest
+import xarray as xr
+from numpy import nan
+
+from scores.continuous import mse, quantile_score
+from scores.continuous.threshold_weighted_impl import (
+    SCORING_FUNCS,
+    _auxiliary_funcs,
+    _g_j_rect,
+    _g_j_trap,
+    _phi_j_prime_rect,
+    _phi_j_prime_trap,
+    _phi_j_rect,
+    _phi_j_trap,
+    threshold_weighted_absolute_error,
+    threshold_weighted_quantile_score,
+    threshold_weighted_score,
+    threshold_weighted_squared_error,
+)
+
+DA_FCST = xr.DataArray(
+    data=[[3.0, 1.0, nan], [-4.0, 0.0, 1.0]],
+    dims=["date", "station"],
+    coords=dict(date=["1", "2"], station=[100, 101, 102]),
+)
+
+DA_OBS = xr.DataArray(
+    data=[[nan, 3.0, 5.0], [-4.0, 10.0, -1.0], [3.0, 2.0, -0.2]],
+    dims=["date", "station"],
+    coords=dict(date=["1", "2", "3"], station=[100, 101, 102]),
+)
+
+DA_X1 = xr.DataArray([nan, -2.0, 1.0, 5.0])
+
+EXP_G_J_RECT1 = xr.DataArray([nan, 0.0, 2.0, 3.0])  # a = -1, b = 2
+
+DA_X2 = xr.DataArray(
+    data=[[nan, -2.0, 1.0, 11.0]],
+    dims=["date", "station"],
+    coords=dict(date=["01"], station=[100, 101, 102, 103]),
+)
+
+DA_A = xr.DataArray(
+    data=[-1.0, 0.0, -1.0, 10.0],
+    dims=["station"],
+    coords=dict(station=[100, 101, 102, 103]),
+)
+
+DA_B = xr.DataArray(
+    data=[2.0, 2.0, 0.0, 12.0],
+    dims=["station"],
+    coords=dict(station=[100, 101, 102, 103]),
+)
+
+EXP_G_J_RECT2 = xr.DataArray(
+    data=[[nan, 0.0, 1.0, 1.0]],
+    dims=["date", "station"],
+    coords=dict(date=["01"], station=[100, 101, 102, 103]),
+)
+
+EXP_PHI_J_RECT1 = xr.DataArray([nan, 0.0, 8.0, 54.0])  # a = -1, b = 2
+
+EXP_PHI_J_RECT2 = xr.DataArray(
+    data=[[nan, 0.0, 6.0, 2.0]],
+    dims=["date", "station"],
+    coords=dict(date=["01"], station=[100, 101, 102, 103]),
+)
+
+EXP_PHI_J_PRIME_RECT1 = 4 * EXP_G_J_RECT1
+
+DA_X3 = xr.DataArray([nan, -3.0, 0.0, 2.0, 5.0, 10.0])
+# s = 5
+# a = -2, b = 1, c = 5, d = 8
+EXP_G_J_TRAP1 = xr.DataArray([nan, 0.0, 2 / 3, 2.5, 7 - 3 / 2, 7.0])
+
+DA_A_TRAP = xr.DataArray(
+    data=[[0, 10]],
+    dims=["date", "station"],
+    coords=dict(date=["01"], station=[100, 101]),
+)
+
+DA_B_TRAP = xr.DataArray(
+    data=[[2, 20]],
+    dims=["date", "station"],
+    coords=dict(date=["01"], station=[100, 101]),
+)
+
+DA_C_TRAP = xr.DataArray(
+    data=[[5, 25]],
[…]
