### 2024-07-12 3b8ce3e
add new scores to api.md and readme
