### 2024-07-12 ef9587d
update fcst test data to test test that broadcasting works
