+++ docs/api.md
@@ -30,6 +30,7 @@
 .. autofunction:: scores.continuous.threshold_weighted_quantile_score
 .. autofunction:: scores.continuous.threshold_weighted_absolute_error
 .. autofunction:: scores.continuous.threshold_weighted_squared_error
+```
 
 ## scores.probability
 ```{eval-rst}
