### 2024-07-12 0991d17
Fixed two errors resulting from a call to a deprecated method which has now been updated
