### 2024-07-14 fbb73e9
Apply suggestions from code review
Co-authored-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
Signed-off-by: Nicholas Loveday <48701367+nicholasloveday@users.noreply.github.com>
