### 2024-07-16 10844d1
fix links to notebook
