+++ docs/included.md
@@ -119,19 +119,19 @@
   - [Wikipedia](https://en.wikipedia.org/wiki/Root-mean-square_deviation)
 * - Threshold Weighted Score
   - [API](api.md#scores.continuous.threshold_weighted_score)
-  - &mdash;
+  - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
   - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
   * - Threshold Weighted Squared Error
   - [API](api.md#scores.continuous.threshold_weighted_squared_error)
-  - &mdash;
+  - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
   - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
   * - Threshold Weighted Absolute Error
   - [API](api.md#scores.continuous.threshold_weighted_absolute_error)
-  - &mdash;
+  - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
   - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
   * - Threshold Weighted Quantile Score
   - [API](api.md#scores.continuous.threshold_quantile_score)
-  - &mdash;
+  - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
   - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
 ```
 
