### 2024-07-16 568fb21
update citation style
