+++ docs/included.md
@@ -117,16 +117,16 @@
   - [API](api.md#scores.continuous.rmse)
   - [Tutorial](project:./tutorials/Root_Mean_Squared_Error.md)
   - [Wikipedia](https://en.wikipedia.org/wiki/Root-mean-square_deviation)
+* - Threshold Weighted Absolute Error
+  - [API](api.md#scores.continuous.threshold_weighted_absolute_error)
+  - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
+  - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
 * - Threshold Weighted Score
   - [API](api.md#scores.continuous.threshold_weighted_score)
   - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
   - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
   * - Threshold Weighted Squared Error
   - [API](api.md#scores.continuous.threshold_weighted_squared_error)
-  - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
-  - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
-  * - Threshold Weighted Absolute Error
-  - [API](api.md#scores.continuous.threshold_weighted_absolute_error)
   - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
   - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
   * - Threshold Weighted Quantile Score
