### 2024-07-16 512dd5b
added pyfunc to funcs mentioned in docstring
