### 2024-07-19 b907d8d
refactored tw_squared_error
