+++ tests/continuous/test_threshold_weighted.py
@@ -11,7 +11,7 @@ from numpy import nan
 
 from scores.continuous import mae, mse, quantile_score
 from scores.continuous.threshold_weighted_impl import (
-    SCORING_FUNCS,
+    # SCORING_FUNCS,
     _auxiliary_funcs,
     _g_j_rect,
     _g_j_trap,
@@ -19,10 +19,10 @@ from scores.continuous.threshold_weighted_impl import (
     _phi_j_prime_trap,
     _phi_j_rect,
     _phi_j_trap,
-    threshold_weighted_absolute_error,
-    threshold_weighted_quantile_score,
-    threshold_weighted_score,
-    threshold_weighted_squared_error,
+    # threshold_weighted_absolute_error,
+    # threshold_weighted_quantile_score,
+    # threshold_weighted_score,
+    tw_squared_error,
 )
 
 DA_FCST = xr.DataArray(
@@ -190,142 +190,130 @@ EXP_MSHL = xr.DataArray(
         "scoring_func",
         "interval_where_one",
         "interval_where_positive",
-        "alpha",
-        "huber_param",
+        "kwargs",
         "error_msg_snippet",
     ),
     [
         (
-            "squared_error",
+            tw_squared_error,
             (0, 2, 4),
             None,
-            0.5,
-            None,
+            {},
             "`interval_where_one` must have length 2",
         ),
         (
-            "squared_error",
+            tw_squared_error,
             (0, 2),
             (-10, 10, 50),
-            0.5,
-            None,
+            {},
             "`interval_where_positive` must be length 2 when not `None`",
         ),
-        ("log_error", (0, 1), None, 0.5, None, "`scoring_func` must be one of:"),
-        ("expectile_score", (0, 1), None, None, None, "`alpha` must be supplied"),
-        (
-            "quantile_score",
-            (0, 1),
-            None,
-            0.0,
-            None,
-            "`alpha` must be strictly between 0 and 1",
-        ),
-        ("huber_loss", (0, 1), None, 0.5, None, "`huber_param` must be supplied"),
-        ("huber_loss", (0, 1), None, None, 0.0, "`huber_param` must be positive"),
-        (
-            "huber_loss",
-            (1, 0),
-            None,
-            None,
-            1,
-            "left endpoint of `interval_where_one` must be strictly less than right endpoint",
-        ),
-        (
-            "huber_loss",
-            (1, 0),
-            (-2, 10),
-            None,
-            1,
-            "left endpoint of `interval_where_one` must be strictly less than right endpoint",
-        ),
-        (
-            "huber_loss",
-            (DA_B, DA_A),
-            None,
-            None,
-            1,
-            "left endpoint of `interval_where_one` must be strictly less than right endpoint",
-        ),
-        (
-            "huber_loss",
-            (DA_B, DA_A),
-            (DA_A - 100, DA_B + 100),
-            None,
-            1,
-            "left endpoint of `interval_where_one` must be strictly less than right endpoint",
-        ),
-        ("huber_loss", (-2, 0), (-np.inf, 6), None, 1, "can only be infinite when"),
-        (
-            "huber_loss",
-            (xr.DataArray([0, 0]), xr.DataArray([1, 1])),
-            (xr.DataArray([-np.inf, -1]), xr.DataArray([2, 2])),
-            None,
-            1,
-            "can only be infinite when",
-        ),
-        ("huber_loss", (-2, 0), (-4, np.inf), None, 1, "can only be infinite when"),
-        (
-            "huber_loss",
-            (xr.DataArray([0, 0]), xr.DataArray([1, 1])),
[…]
