+++ src/scores/continuous/threshold_weighted_impl.py
@@ -18,207 +18,209 @@ from scores.continuous.consistent_impl import (
 )
 from scores.typing import FlexibleDimensionTypes
 
-HUBER_FUNCS = ["huber_loss", "scaled_huber_loss"]
-ALPHA_FUNCS = ["quantile_score", "expectile_score"]
-OTHER_FUNCS = ["squared_error", "absolute_error"]
-SCORING_FUNCS = OTHER_FUNCS + ALPHA_FUNCS + HUBER_FUNCS
+# HUBER_FUNCS = ["huber_loss", "scaled_huber_loss"]
+# ALPHA_FUNCS = ["quantile_score", "expectile_score"]
+# OTHER_FUNCS = ["squared_error", "absolute_error"]
+# SCORING_FUNCS = OTHER_FUNCS + ALPHA_FUNCS + HUBER_FUNCS
 AuxFuncType = Callable[[xr.DataArray], xr.DataArray]
 EndpointType = Union[int, float, xr.DataArray]
 
 
-def threshold_weighted_score(
-    fcst: xr.DataArray,
-    obs: xr.DataArray,
-    scoring_func: Literal[
-        "squared_error",
-        "absolute_error",
-        "quantile_score",
-        "expectile_score",
-        "huber_loss",
-        "scaled_huber_loss",
-    ],
-    interval_where_one: Tuple[EndpointType, EndpointType],
-    *,
-    interval_where_positive: Optional[Tuple[EndpointType, EndpointType]] = None,
-    alpha: Optional[float] = None,
-    huber_param: Optional[float] = None,
-    reduce_dims: Optional[FlexibleDimensionTypes] = None,
-    preserve_dims: Optional[FlexibleDimensionTypes] = None,
-    weights: Optional[xr.DataArray] = None,
-) -> xr.DataArray:
-    """
-    Returns the scores calculated using the specified threshold weighted scoring function,
-    averaged over the specified dimensions in `dims`.
+# def threshold_weighted_score(
+#     fcst: xr.DataArray,
+#     obs: xr.DataArray,
+#     scoring_func: Literal[
+#         "squared_error",
+#         "absolute_error",
+#         "quantile_score",
+#         "expectile_score",
+#         "huber_loss",
+#         "scaled_huber_loss",
+#     ],
+#     interval_where_one: Tuple[EndpointType, EndpointType],
+#     *,
+#     interval_where_positive: Optional[Tuple[EndpointType, EndpointType]] = None,
+#     alpha: Optional[float] = None,
+#     huber_param: Optional[float] = None,
+#     reduce_dims: Optional[FlexibleDimensionTypes] = None,
+#     preserve_dims: Optional[FlexibleDimensionTypes] = None,
+#     weights: Optional[xr.DataArray] = None,
+# ) -> xr.DataArray:
+#     """
+#     Returns the scores calculated using the specified threshold weighted scoring function,
+#     averaged over the specified dimensions in `dims`.
 
-    Taggart (2022) showed that many commonly used scoring functions, such as the squared
-    error and absolute error scoring functions, can be decomposed as a sum of weighted
-    scoring functions, where the weights sum to 1, and where the original scoring function and
-    each member of its decomposition are consistent for the same statistical functional.
-    Each member of the decomposition measures predictive performance with emphasis on the
-    region specified by the corresponding weight.
+#     Taggart (2022) showed that many commonly used scoring functions, such as the squared
+#     error and absolute error scoring functions, can be decomposed as a sum of weighted
+#     scoring functions, where the weights sum to 1, and where the original scoring function and
+#     each member of its decomposition are consistent for the same statistical functional.
+#     Each member of the decomposition measures predictive performance with emphasis on the
+#     region specified by the corresponding weight.
 
[…]
