+++ src/scores/continuous/__init__.py
@@ -23,10 +23,10 @@ from scores.continuous.standard_impl import (
     rmse,
 )
 from scores.continuous.threshold_weighted_impl import (
-    threshold_weighted_absolute_error,
-    threshold_weighted_quantile_score,
-    threshold_weighted_score,
-    threshold_weighted_squared_error,
+    # threshold_weighted_absolute_error,
+    # threshold_weighted_quantile_score,
+    # threshold_weighted_score,
+    tw_squared_error,
 )
 from scores.processing.isoreg_impl import isotonic_fit
 
@@ -47,8 +47,8 @@ __all__ = [
     "consistent_expectile_score",
     "consistent_huber_score",
     "consistent_quantile_score",
-    "threshold_weighted_score",
-    "threshold_weighted_quantile_score",
-    "threshold_weighted_absolute_error",
-    "threshold_weighted_squared_error",
+    # "threshold_weighted_score",
+    # "threshold_weighted_quantile_score",
+    # "threshold_weighted_absolute_error",
+    "tw_squared_error",
 ]
