+++ src/scores/continuous/threshold_weighted_impl.py
@@ -503,6 +503,10 @@ def tw_squared_error(
         interval_where_one: endpoints of the interval where the weights are 1.
             Must be increasing. Infinite endpoints are permissible. By supplying a tuple of
             arrays, endpoints can vary with dimension.
+        interval_where_positive: endpoints of the interval where the weights are positive.
+            Must be increasing. Infinite endpoints are only permissible when the corresponding
+            `interval_where_one` endpoint is infinite. By supplying a tuple of
+            arrays, endpoints can vary with dimension.
         reduce_dims: Optionally specify which dimensions to reduce when
             calculating the threshold_weighted_square_error. All other dimensions will be preserved. As a
             special case, 'all' will allow all dimensions to be reduced. Only one
@@ -539,67 +543,83 @@ def tw_squared_error(
     )
 
 
-# def threshold_weighted_absolute_error(
-#     fcst: xr.DataArray,
-#     obs: xr.DataArray,
-#     interval_where_one: Tuple[EndpointType, EndpointType],
-#     reduce_dims: Optional[FlexibleDimensionTypes] = None,
-#     preserve_dims: Optional[FlexibleDimensionTypes] = None,
-#     weights: Optional[xr.DataArray] = None,
-# ) -> xr.DataArray:
-#     """
-#     Returns the scores computed using a threshold weighted scoring
-#     function derived from the absolute error function.
+def tw_absolute_error(
+    fcst: xr.DataArray,
+    obs: xr.DataArray,
+    interval_where_one: Tuple[EndpointType, EndpointType],
+    *,
+    interval_where_positive: Optional[Tuple[EndpointType, EndpointType]] = None,
+    reduce_dims: Optional[FlexibleDimensionTypes] = None,
+    preserve_dims: Optional[FlexibleDimensionTypes] = None,
+    weights: Optional[xr.DataArray] = None,
+) -> xr.DataArray:
+    """
+    Returns the threshold weighted absolute error function.
 
-#     This is a convenience function for :py:func:`threshold_weighted_score`
-#     with the option `scoring_func="absolute_error"`. The weight is
-#     1 on the specified interval and 0 elsewhere. For more flexible weighting schemes,
-#     see :py:func:`scores.continuous.threshold_weighted_score` and
-#     :py:func:`scores.continuous.consistent_absolute_error`.
+    For more flexible weighting schemes,
+    see :py:func:`scores.continuous.consistent_absolute_error`.
 
-#     Args:
-#         fcst: array of forecast values.
-#         obs: array of corresponding observation values.
-#         interval_where_one: endpoints of the interval where the weights are 1.
-#             Must be increasing. Infinite endpoints are permissible. By supplying a tuple of
-#             arrays, endpoints can vary with dimension.
-#         reduce_dims: Optionally specify which dimensions to reduce when
-#             calculating the threshold_weighted_absolute_error. All other dimensions will be preserved. As a
-#             special case, 'all' will allow all dimensions to be reduced. Only one
-#             of `reduce_dims` and `preserve_dims` can be supplied. The default behaviour
-#             if neither are supplied is to reduce all dims.
-#         preserve_dims: Optionally specify which dimensions to preserve when calculating
-#             the threshold_weighted_absolute_error. All other dimensions will be reduced. As a special case, 'all'
-#             will allow all dimensions to be preserved. In this case, the result will be in
-#             the same shape/dimensionality as the forecast, and the errors will be the threshold_weighted_absolute_error
-#             at each point (i.e. single-value comparison against observed), and the
-#             forecast and observed dimensions must match precisely. Only one of `reduce_dims`
-#             and `preserve_dims` can be supplied. The default behaviour if neither are supplied
-#             is to reduce all dims.
[…]
