+++ tests/continuous/test_threshold_weighted.py
@@ -20,7 +20,7 @@ from scores.continuous.threshold_weighted_impl import (
     _phi_j_rect,
     _phi_j_trap,
     tw_absolute_error,
-    # threshold_weighted_quantile_score,
+    tw_quantile_score,
     # threshold_weighted_score,
     tw_squared_error,
 )
@@ -449,8 +449,13 @@ def test__auxiliary_funcs2(interval_where_one, interval_where_positive, a, b, c,
 @pytest.mark.parametrize(
     ("scoring_func", "kwargs", "expected"),
     [
-        (tw_squared_error, {}, mse(DA_FCST1, DA_OBS1, preserve_dims="all")),
-        (tw_absolute_error, {}, mae(DA_FCST1, DA_OBS1, preserve_dims="all")),
+        (tw_squared_error, {}, mse(DA_FCST1, DA_OBS1, preserve_dims=["date", "station"])),
+        (tw_absolute_error, {}, mae(DA_FCST1, DA_OBS1, preserve_dims=["date", "station"])),
+        (
+            tw_quantile_score,
+            {"alpha": 0.3},
+            quantile_score(DA_FCST1, DA_OBS1, alpha=0.3, preserve_dims=["date", "station"]),
+        ),
     ],
 )
 def test_threshold_weighted_scores1(scoring_func, kwargs, expected):
@@ -465,7 +470,9 @@ def test_threshold_weighted_scores1(scoring_func, kwargs, expected):
     xr.testing.assert_allclose(result, expected)
 
 
-@pytest.mark.parametrize(("scoring_func", "kwargs"), [(tw_squared_error, {}), (tw_absolute_error, {})])
+@pytest.mark.parametrize(
+    ("scoring_func", "kwargs"), [(tw_squared_error, {}), (tw_absolute_error, {}), (tw_quantile_score, {"alpha": 0.3})]
+)
 def test_threshold_weighted_scores2(scoring_func, kwargs):
     """
     Tests that each threshold weight score gives results as expected.
@@ -473,16 +480,17 @@ def test_threshold_weighted_scores2(scoring_func, kwargs):
     sum to 1 is the same as the score with a weight of 1 everywhere.
     Tests for rectangular weights.
     """
-
     kwargs["preserve_dims"] = ["date", "station"]
 
-    score1 = scoring_func(DA_FCST1, DA_OBS1, (-np.inf, 0), **kwargs)
-    score2 = scoring_func(DA_FCST1, DA_OBS1, (0, np.inf), **kwargs)
-    score = scoring_func(DA_FCST1, DA_OBS1, (-np.inf, np.inf), **kwargs)
+    score1 = scoring_func(DA_FCST1, DA_OBS1, interval_where_one=(-np.inf, 0), **kwargs)
+    score2 = scoring_func(DA_FCST1, DA_OBS1, interval_where_one=(0, np.inf), **kwargs)
+    score = scoring_func(DA_FCST1, DA_OBS1, interval_where_one=(-np.inf, np.inf), **kwargs)
     xr.testing.assert_allclose(score1 + score2, score)
 
 
-@pytest.mark.parametrize(("scoring_func", "kwargs"), [(tw_squared_error, {}), (tw_absolute_error, {})])
+@pytest.mark.parametrize(
+    ("scoring_func", "kwargs"), [(tw_squared_error, {}), (tw_absolute_error, {}), (tw_quantile_score, {"alpha": 0.3})]
+)
 def test_threshold_weighted_scores3(scoring_func, kwargs):
     """
     Tests that each threshold weight score gives results as expected.
[…]
