+++ src/scores/continuous/threshold_weighted_impl.py
@@ -554,10 +554,10 @@ def tw_absolute_error(
     weights: Optional[xr.DataArray] = None,
 ) -> xr.DataArray:
     """
-    Returns the threshold weighted absolute error function.
+    Returns the threshold weighted absolute error.
 
     For more flexible weighting schemes,
-    see :py:func:`scores.continuous.consistent_absolute_error`.
+    see :py:func:`scores.continuous.consistent_quantile_score`.
 
     Two types of threshold weighting are supported: rectangular and trapezoidal.
 
@@ -622,68 +622,84 @@ def tw_absolute_error(
     )
 
 
-# def threshold_weighted_quantile_score(
-#     fcst: xr.DataArray,
-#     obs: xr.DataArray,
-#     interval_where_one: Tuple[EndpointType, EndpointType],
-#     alpha: float,
-#     reduce_dims: Optional[FlexibleDimensionTypes] = None,
-#     preserve_dims: Optional[FlexibleDimensionTypes] = None,
-#     weights: Optional[xr.DataArray] = None,
-# ) -> xr.DataArray:
-#     """
-#     Returns the scores computed using a threshold weighted scoring
-#     function derived from the standard `quantile_score`.
+def tw_quantile_score(
+    fcst: xr.DataArray,
+    obs: xr.DataArray,
+    alpha: float,
+    interval_where_one: Tuple[EndpointType, EndpointType],
+    *,
+    interval_where_positive: Optional[Tuple[EndpointType, EndpointType]] = None,
+    reduce_dims: Optional[FlexibleDimensionTypes] = None,
+    preserve_dims: Optional[FlexibleDimensionTypes] = None,
+    weights: Optional[xr.DataArray] = None,
+) -> xr.DataArray:
+    """
+    Returns the threshold weighted quantile score.
 
-#     This is a convenience function for :py:func:`threshold_weighted_score`
-#     with the option `scoring_func="quantile_score"`. The weight is
-#     1 on the specified interval and 0 elsewhere. For more flexible weighting schemes,
-#     see :py:func:`scores.continuous.threshold_weighted_score` and
-#     :py:func:`scores.continuous.consistent_quantile_score`.
+    For more flexible weighting schemes,
+    see :py:func:`scores.continuous.consistent_quantile_score`.
 
-#     Args:
-#         fcst: array of forecast values.
-#         obs: array of corresponding observation values.
-#         interval_where_one: endpoints of the interval where the weights are 1.
-#             Must be increasing. Infinite endpoints are permissible. By supplying a tuple of
-#             arrays, endpoints can vary with dimension.
-#         alpha: the quantile level.
-#         reduce_dims: Optionally specify which dimensions to reduce when
-#             calculating the threshold_weighted_quantile_score. All other dimensions will be preserved. As a
-#             special case, 'all' will allow all dimensions to be reduced. Only one
-#             of `reduce_dims` and `preserve_dims` can be supplied. The default behaviour
-#             if neither are supplied is to reduce all dims.
-#         preserve_dims: Optionally specify which dimensions to preserve when calculating
-#             the threshold_weighted_quantile_score. All other dimensions will be reduced. As a special case, 'all'
-#             will allow all dimensions to be preserved. In this case, the result will be in
-#             the same shape/dimensionality as the forecast, and the errors will be the threshold_weighted_quantile_score
-#             at each point (i.e. single-value comparison against observed), and the
-#             forecast and observed dimensions must match precisely. Only one of `reduce_dims`
-#             and `preserve_dims` can be supplied. The default behaviour if neither are supplied
-#             is to reduce all dims.
-#         weights: Optionally provide an array for weighted averaging (e.g. by area, by latitude,
-#             by population, custom)
+    Two types of threshold weighting are supported: rectangular and trapezoidal.
 
-#     Returns:
[…]
