### 2024-07-22 f7bfa7d
add checks for alpha values
