+++ tests/continuous/test_threshold_weighted.py
@@ -21,7 +21,7 @@ from scores.continuous.threshold_weighted_impl import (
     _phi_j_trap,
     tw_absolute_error,
     tw_quantile_score,
-    # threshold_weighted_score,
+    tw_expectile_score,
     tw_squared_error,
 )
 
@@ -209,6 +209,48 @@ EXP_MSHL = xr.DataArray(
             "`interval_where_positive` must be length 2 when not `None`",
         ),
         (
+            tw_absolute_error,
+            (0, 2, 4),
+            None,
+            {},
+            "`interval_where_one` must have length 2",
+        ),
+        (
+            tw_absolute_error,
+            (0, 2),
+            (-10, 10, 50),
+            {},
+            "`interval_where_positive` must be length 2 when not `None`",
+        ),
+        (
+            tw_quantile_score,
+            (0, 2, 4),
+            None,
+            {"alpha": 0.3},
+            "`interval_where_one` must have length 2",
+        ),
+        (
+            tw_quantile_score,
+            (0, 2),
+            (-10, 10, 50),
+            {"alpha": 0.3},
+            "`interval_where_positive` must be length 2 when not `None`",
+        ),
+        (
+            tw_expectile_score,
+            (0, 2, 4),
+            None,
+            {"alpha": 0.3},
+            "`interval_where_one` must have length 2",
+        ),
+        (
+            tw_expectile_score,
+            (0, 2),
+            (-10, 10, 50),
+            {"alpha": 0.3},
+            "`interval_where_positive` must be length 2 when not `None`",
+        ),
+        (
             tw_quantile_score,
             (0, 1),
             None,
@@ -222,8 +264,21 @@ EXP_MSHL = xr.DataArray(
             {"alpha": 1},
             "`alpha` must be strictly between 0 and 1",
         ),
+        (
+            tw_expectile_score,
+            (0, 1),
+            None,
+            {"alpha": 0},
+            "`alpha` must be strictly between 0 and 1",
+        ),
+        (
+            tw_expectile_score,
+            (0, 1),
+            None,
+            {"alpha": 1},
+            "`alpha` must be strictly between 0 and 1",
+        ),
         #         ("log_error", (0, 1), None, 0.5, None, "`scoring_func` must be one of:"),
-        #         ("expectile_score", (0, 1), None, None, None, "`alpha` must be supplied"),
         #         ("huber_loss", (0, 1), None, 0.5, None, "`huber_param` must be supplied"),
         #         ("huber_loss", (0, 1), None, None, 0.0, "`huber_param` must be positive"),
         #         (
@@ -468,6 +523,11 @@ def test__auxiliary_funcs2(interval_where_one, interval_where_positive, a, b, c,
             {"alpha": 0.3},
             quantile_score(DA_FCST1, DA_OBS1, alpha=0.3, preserve_dims=["date", "station"]),
         ),
+        (
+            tw_expectile_score,
+            {"alpha": 0.5},
+            mse(DA_FCST1, DA_OBS1, preserve_dims=["date", "station"]),
+        ),
     ],
 )
 def test_threshold_weighted_scores1(scoring_func, kwargs, expected):
@@ -483,7 +543,13 @@ def test_threshold_weighted_scores1(scoring_func, kwargs, expected):
 
 
 @pytest.mark.parametrize(
[…]
