### 2024-07-22 6d48abc
add scores to readme
