### 2024-07-22 96b0581
add tw scores to api
