### 2024-07-22 cffe04e
add comments about scaling of phi
