### 2024-07-22 be7f543
update docstrings
