### 2024-07-22 c6d8f67
remove chaining of wheres
