+++ src/scores/continuous/threshold_weighted_impl.py
@@ -4,7 +4,7 @@ See Taggart (2022) https://doi.org/10.1002/qj.4206
 """
 
 import functools
-from typing import Callable, Literal, Optional, Tuple, Union
+from typing import Callable, Optional, Tuple, Union
 
 import numpy as np
 import xarray as xr
@@ -27,8 +27,8 @@ def _check_tws_args(
     interval_where_positive: Optional[Tuple[EndpointType, EndpointType]],
 ):
     """
-    Some argument checks for `threshold_weighted_score`.
-    Checks for valid interval endpoints are done in `_auxiliary_funcs`.
+    Some argument checks for the threshold weighted scores.
+    Checks for valid interval endpoints are done in :py:func`_auxiliary_funcs`.
 
     Args:
         interval_where_one: endpoints of the interval where the weights are 1.
@@ -36,12 +36,12 @@ def _check_tws_args(
             arrays, endpoints can vary with dimension.
         interval_where_positive: endpoints of the interval where the weights are positive.
             Must be increasing. Infinite endpoints are only permissible when the corresponding
-            `interval_where_one` endpoint is infinite. By supplying a tuple of
+            ``interval_where_one`` endpoint is infinite. By supplying a tuple of
             arrays, endpoints can vary with dimension.
 
     Raises:
-        ValueError: if `interval_where_one` is not length 2.
-        ValueError: if `interval_where_one` is not increasing.
+        ValueError: if ``interval_where_one`` is not length 2.
+        ValueError: if ``interval_where_one`` is not increasing.
     """
     if len(interval_where_one) != 2:
         raise ValueError("`interval_where_one` must have length 2")
@@ -69,14 +69,14 @@ def _auxiliary_funcs(
             arrays, endpoints can vary with dimension.
         interval_where_positive: endpoints of the interval where the weights are positive.
             Must be increasing. Infinite endpoints are only permissible when the corresponding
-            `interval_where_one` endpoint is infinite. By supplying a tuple of
+            ``interval_where_one`` endpoint is infinite. By supplying a tuple of
             arrays, endpoints can vary with dimension.
 
     Raises:
-        ValueError: if the left endpoint of `interval_where_one` is not less than
+        ValueError: if the left endpoint of ``interval_where_one`` is not less than
             the right endpoint.
-        ValueError: if an `interval_where_positive` endpoint is infinite when the
-            `interval_where_one` endpoint is infinite.
+        ValueError: if an ``interval_where_positive`` endpoint is infinite when the
+            ``interval_where_one`` endpoint is infinite.
         ValueError: If the right endpoint of `interval_where_positive` is not greater
             than the right endpoint of `interval_where_one` and neither are infinite.
     """
@@ -152,8 +152,8 @@ def _g_j_rect(a: EndpointType, b: EndpointType, x: xr.DataArray) -> xr.DataArray
     The formula is based on the first row of Table B1 from Taggart (2022).
 
     Args:
-        a: left endpoint of interval where weight = 1. Can be `-np.inf`.
-        b: right endpoint of the interval where weight = 1. Can be `np.inf`.
+        a: left endpoint of interval where weight = 1. Can be ``-np.inf``.
+        b: right endpoint of the interval where weight = 1. Can be ``np.inf``.
         x: points where g_j is to be evaluated.
 
     Returns:
@@ -181,8 +181,8 @@ def _phi_j_rect(a: EndpointType, b: EndpointType, x: xr.DataArray) -> xr.DataArr
     The formula is based on the second row of Table B1 from Taggart (2022).
 
     Args:
-        a: left endpoint of interval where weight = 1. Can be `-np.inf`.
-        b: right endpoint of the interval where weight = 1. Can be `np.inf`.
[…]
