+++ docs/included.md
@@ -129,12 +129,12 @@
   - [API](api.md#scores.continuous.tw_huber_loss)
   - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
   - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
-* - Threshold Weighted Squared Error
-  - [API](api.md#scores.continuous.tw_squared_error)
-  - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
-  - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
 * - Threshold Weighted Quantile Score
   - [API](api.md#scores.continuous.tw_quantile_score)
+  - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
+  - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
+* - Threshold Weighted Squared Error
+  - [API](api.md#scores.continuous.tw_squared_error)
   - [Tutorial](project:./tutorials/Threshold_Weighted_Scores.md)
   - [Taggart (2022)](https://doi.org/10.1002/qj.4206)
 ```
