### 2024-07-24 f55504e
Update threshold weighted notebook such that the plots render in the built docs Ignore erroneous mypy warnings where mypy is incorrect Add a missing backtick to a docstring
