### 2024-07-24 b0f5ba6
Update the maintainer requirements to include ipywidgets to improve plotly rendering for readthedocs Update the maintainer notes with a handy reference command for working with forks
