### 2024-07-31 81854c5
Add journal article link early in README
