+++ README.md
@@ -7,9 +7,10 @@
 
 `scores` is a Python package containing mathematical functions for the verification, evaluation and optimisation of forecasts, predictions or models. It supports labelled n-dimensional (multidimensional) data, which is used in many scientific fields and in machine learning. At present, `scores` primarily supports the geoscience communities; in particular, the meteorological, climatological and oceanographic communities.
 
-Documentation is hosted at [scores.readthedocs.io](https://scores.readthedocs.io).  
-Source code is hosted at [github.com/nci/scores](https://github.com/nci/scores).  
-The tutorial gallery is hosted at [as part of the documentation, here](https://scores.readthedocs.io/en/stable/tutorials/Tutorial_Gallery.html). 
+Documentation is hosted at [scores.readthedocs.io](https://scores.readthedocs.io)  
+Source code is hosted at [github.com/nci/scores](https://github.com/nci/scores)  
+The tutorial gallery is hosted at [as part of the documentation, here](https://scores.readthedocs.io/en/stable/tutorials/Tutorial_Gallery.html)  
+Journal article - [*scores: A Python package for verifying and evaluating models and predictions with xarray*](https://doi.org/10.21105/joss.06889)
 
 ## Overview
 Here is a **curated selection** of the metrics, tools and statistical tests included in `scores`:
