### 2024-07-31 657f114
Different options for mentioning journal article at start of README
