### 2024-07-31 5608de5
Modify trapezoid call to work with either numpy 1 or 2 seamlessly
