### 2024-07-31 nci/scores#554 @tennlee
I've now fixed everything in the tests where the root cause can be addressed in the scores code. There is one warning which are raised deep in libraries calling other libraries. As maintainer, I find it helpful to keep seeing these in the test results rather than wrapping and hiding them where they might pass silently and be neglected or forgotten over time. Addressing the issues under our control has helped the code quality, and I think I can now close this issue. Thanks for raising it!
