+++ README.md
@@ -9,12 +9,12 @@
 
 Documentation is hosted at [scores.readthedocs.io](https://scores.readthedocs.io)  
 Source code is hosted at [github.com/nci/scores](https://github.com/nci/scores)  
-The tutorial gallery is hosted at [as part of the documentation, here](https://scores.readthedocs.io/en/stable/tutorials/Tutorial_Gallery.html)  
+The tutorial gallery is hosted [as part of the documentation, here](https://scores.readthedocs.io/en/stable/tutorials/Tutorial_Gallery.html)  
 Journal article - [*scores: A Python package for verifying and evaluating models and predictions with xarray*](https://doi.org/10.21105/joss.06889)
 
 Documentation is hosted at [scores.readthedocs.io](https://scores.readthedocs.io)  
 Source code is hosted at [github.com/nci/scores](https://github.com/nci/scores)  
-The tutorial gallery is hosted at [as part of the documentation, here](https://scores.readthedocs.io/en/stable/tutorials/Tutorial_Gallery.html)  
+The tutorial gallery is hosted [as part of the documentation, here](https://scores.readthedocs.io/en/stable/tutorials/Tutorial_Gallery.html)  
 Journal article is available at [https://doi.org/10.21105/joss.06889](https://doi.org/10.21105/joss.06889)
 
 Documentation: [scores.readthedocs.io](https://scores.readthedocs.io)  
