### 2024-07-31 4a668c7
Slightly update docstring to improve rendering on readthedocs
