+++ src/scores/continuous/quantile_loss_impl.py
@@ -61,7 +61,6 @@ def quantile_score(
            (1-\\alpha) x & x > 0\\end{cases}
 
         where:
-
             - :math:`\\alpha` is the targeted quantile.
             - :math:`x` is the difference, fcst - obs
 
