### 2024-08-01 97b227f
improved docstrings when talking about threshold weighting
