+++ src/scores/continuous/threshold_weighted_impl.py
@@ -31,10 +31,10 @@ def _check_tws_args(
     Checks for valid interval endpoints are done in :py:func`_auxiliary_funcs`.
 
     Args:
-        interval_where_one: endpoints of the interval where the weights are 1.
+        interval_where_one: endpoints of the interval where the threshold weights are 1.
             Must be increasing. Infinite endpoints are permissible. By supplying a tuple of
             arrays, endpoints can vary with dimension.
-        interval_where_positive: endpoints of the interval where the weights are positive.
+        interval_where_positive: endpoints of the interval where the threshold weights are positive.
             Must be increasing. Infinite endpoints are only permissible when the corresponding
             ``interval_where_one`` endpoint is infinite. By supplying a tuple of
             arrays, endpoints can vary with dimension.
@@ -64,10 +64,10 @@ def _auxiliary_funcs(
     Args:
         fcst: array of forecast values.
         obs: array of corresponding observation values.
-        interval_where_one: endpoints of the interval where the weights are 1.
+        interval_where_one: endpoints of the interval where the threshold weights are 1.
             Must be increasing. Infinite endpoints are permissible. By supplying a tuple of
             arrays, endpoints can vary with dimension.
-        interval_where_positive: endpoints of the interval where the weights are positive.
+        interval_where_positive: endpoints of the interval where the threshold weights are positive.
             Must be increasing. Infinite endpoints are only permissible when the corresponding
             ``interval_where_one`` endpoint is infinite. By supplying a tuple of
             arrays, endpoints can vary with dimension.
@@ -81,7 +81,7 @@ def _auxiliary_funcs(
             than the right endpoint of ``interval_where_one`` and neither are infinite.
     """
 
-    if interval_where_positive is None:  # rectangular weight
+    if interval_where_positive is None:  # rectangular threshold weight
         a, b = interval_where_one
 
         if isinstance(a, (float, int)):
@@ -99,7 +99,7 @@ def _auxiliary_funcs(
         phi = functools.partial(_phi_j_rect, a, b)
         phi_prime = functools.partial(_phi_j_prime_rect, a, b)
 
-    else:  # trapezoidal weight
+    else:  # trapezoidal threshold weight
         a, d = interval_where_positive
         b, c = interval_where_one
 
@@ -148,12 +148,12 @@ def _auxiliary_funcs(
 def _g_j_rect(a: EndpointType, b: EndpointType, x: xr.DataArray) -> xr.DataArray:
     """
     Returns values of a nondecreasing function g_j, where g_j is obtained by integrating
-    a rectangular weight function. The weight is 1 on the interval [a, b) and 0 otherwise.
+    a rectangular threshold weight function. The threshold weight is 1 on the interval [a, b) and 0 otherwise.
     The formula is based on the first row of Table B1 from Taggart (2022).
 
     Args:
-        a: left endpoint of interval where weight = 1. Can be ``-np.inf``.
-        b: right endpoint of the interval where weight = 1. Can be ``np.inf``.
+        a: left endpoint of interval where the threshold weight = 1. Can be ``-np.inf``.
+        b: right endpoint of the interval where the threshold weight = 1. Can be ``np.inf``.
         x: points where g_j is to be evaluated.
 
     Returns:
@@ -177,12 +177,13 @@ def _g_j_rect(a: EndpointType, b: EndpointType, x: xr.DataArray) -> xr.DataArray
 def _phi_j_rect(a: EndpointType, b: EndpointType, x: xr.DataArray) -> xr.DataArray:
     """
     Returns values of a convex function phi_j, where phi_j is obtained by integrating
-    a rectangular weight function. The weight is 1 on the interval [a, b) and 0 otherwise.
+    a rectangular threshold weight function. The threshold weight is 1 on the interval
+    [a, b) and 0 otherwise.
[…]
