### 2024-08-01 56d9aa7
Remove nbviewer note from maintainer's guide now that nbviewer has been removed from readthedocs
