+++ README.md
@@ -17,6 +17,11 @@ Source code is hosted at [github.com/nci/scores](https://github.com/nci/scores)
 The tutorial gallery is hosted [as part of the documentation, here](https://scores.readthedocs.io/en/stable/tutorials/Tutorial_Gallery.html)  
 Journal article is available at [https://doi.org/10.21105/joss.06889](https://doi.org/10.21105/joss.06889)
 
+Documentation is hosted at [scores.readthedocs.io](https://scores.readthedocs.io)  
+Source code is hosted at [github.com/nci/scores](https://github.com/nci/scores)  
+The tutorial gallery is hosted [as part of the documentation, here](https://scores.readthedocs.io/en/stable/tutorials/Tutorial_Gallery.html)  
+Journal article - [Leeuwenburg et al. (2024)](https://doi.org/10.21105/joss.06889)
+
 Documentation: [scores.readthedocs.io](https://scores.readthedocs.io)  
 Source code: [github.com/nci/scores](https://github.com/nci/scores)  
 Tutorial gallery: [available here](https://scores.readthedocs.io/en/stable/tutorials/Tutorial_Gallery.html)  
