### 2024-08-02 3f7f7a9
608 add threshold weighted scores (#609)
This is a major new feature introducing threshold-weighted consistent scoring to the API
* add new scores to api.md and readme
* update fcst test data to test test that broadcasting works
* update docstrings
* produced tw scoring tutorial
* added to_absolute_error
* added tw_quantile_score
* add checks for alpha values
* added tw expectile score
* add comments about scaling of phi
* Ignore erroneous mypy warnings where mypy is incorrect

---------

Signed-off-by: Nicholas Loveday <48701367+nicholasloveday@users.noreply.github.com>
Co-authored-by: Stepha
[…]
