+++ src/scores/continuous/threshold_weighted_impl.py
@@ -1,0 +1,739 @@
+"""
+Implementation of some basic threshold weighted scoring functions.
+See Taggart (2022) https://doi.org/10.1002/qj.4206
+"""
+
+import functools
+from typing import Callable, Optional, Tuple, Union
+
+import numpy as np
+import xarray as xr
+
+from scores.continuous.consistent_impl import (
+    check_alpha,
+    check_huber_param,
+    consistent_expectile_score,
+    consistent_huber_score,
+    consistent_quantile_score,
+)
+from scores.typing import FlexibleDimensionTypes
+
+AuxFuncType = Callable[[xr.DataArray], xr.DataArray]
+EndpointType = Union[int, float, xr.DataArray]
+
+
+def _check_tws_args(
+    interval_where_one: Tuple[EndpointType, EndpointType],
+    interval_where_positive: Optional[Tuple[EndpointType, EndpointType]],
+):
+    """
+    Some argument checks for the threshold weighted scores.
+    Checks for valid interval endpoints are done in :py:func`_auxiliary_funcs`.
+
+    Args:
+        interval_where_one: endpoints of the interval where the threshold weights are 1.
+            Must be increasing. Infinite endpoints are permissible. By supplying a tuple of
+            arrays, endpoints can vary with dimension.
+        interval_where_positive: endpoints of the interval where the threshold weights are positive.
+            Must be increasing. Infinite endpoints are only permissible when the corresponding
+            ``interval_where_one`` endpoint is infinite. By supplying a tuple of
+            arrays, endpoints can vary with dimension.
+
+    Raises:
+        ValueError: if ``interval_where_one`` is not length 2.
+        ValueError: if ``interval_where_one`` is not increasing.
+    """
+    if len(interval_where_one) != 2:
+        raise ValueError("`interval_where_one` must have length 2")
+
+    if interval_where_positive is not None and len(interval_where_positive) != 2:
+        raise ValueError("`interval_where_positive` must be length 2 when not `None`")
+
+
+def _auxiliary_funcs(
+    fcst: xr.DataArray,
+    obs: xr.DataArray,
+    interval_where_one: Tuple[EndpointType, EndpointType],
+    interval_where_positive: Optional[Tuple[EndpointType, EndpointType]],
+) -> Tuple[AuxFuncType, AuxFuncType, AuxFuncType]:
+    """
+    Returns the three auxiliary functions g, phi and phi_prime
+    which are used to construct quantile, expectile or Huber mean scoring functions.
+    See Equations (8), (10) and (11) from Taggart (2022) for the role of g, phi and phi_prime.
+
+    Args:
+        fcst: array of forecast values.
+        obs: array of corresponding observation values.
+        interval_where_one: endpoints of the interval where the threshold weights are 1.
+            Must be increasing. Infinite endpoints are permissible. By supplying a tuple of
+            arrays, endpoints can vary with dimension.
+        interval_where_positive: endpoints of the interval where the threshold weights are positive.
+            Must be increasing. Infinite endpoints are only permissible when the corresponding
+            ``interval_where_one`` endpoint is infinite. By supplying a tuple of
+            arrays, endpoints can vary with dimension.
+
+    Raises:
+        ValueError: if the left endpoint of ``interval_where_one`` is not less than
+            the right endpoint.
+        ValueError: if an ``interval_where_positive`` endpoint is infinite when the
+            ``interval_where_one`` endpoint is infinite.
+        ValueError: If the right endpoint of ``interval_where_positive`` is not greater
[…]
