+++ tests/continuous/test_threshold_weighted.py
@@ -1,0 +1,744 @@
+# pylint: disable=use-dict-literal
+"""
+Tests for scores.continuous.threshold_weighted_impl
+"""
+
+import dask
+import numpy as np
+import pytest
+import xarray as xr
+from numpy import nan
+
+from scores.continuous import mae, mse, quantile_score
+from scores.continuous.threshold_weighted_impl import (
+    _auxiliary_funcs,
+    _g_j_rect,
+    _g_j_trap,
+    _phi_j_prime_rect,
+    _phi_j_prime_trap,
+    _phi_j_rect,
+    _phi_j_trap,
+    tw_absolute_error,
+    tw_expectile_score,
+    tw_huber_loss,
+    tw_quantile_score,
+    tw_squared_error,
+)
+
+DA_FCST = xr.DataArray(
+    data=[[[3.0, 1.0, nan, 2], [3.0, 1.0, nan, 2]], [[-4.0, 0.0, 1.0, 2], [-4.0, 0.0, 1.0, 2]]],
+    dims=["date", "lead_day", "station"],
+    coords=dict(
+        date=["1", "2"],
+        lead_day=[1, 1],
+        station=[100, 101, 102, 0],
+    ),
+)
+
+DA_OBS = xr.DataArray(
+    data=[[nan, 3.0, 5.0], [-4.0, 10.0, -1.0], [3.0, 2.0, -0.2]],
+    dims=["date", "station"],
+    coords=dict(date=["1", "2", "3"], station=[100, 101, 102]),
+)
+
+DA_X1 = xr.DataArray([nan, -2.0, 1.0, 5.0])
+
+EXP_G_J_RECT1 = xr.DataArray([nan, 0.0, 2.0, 3.0])  # a = -1, b = 2
+
+DA_X2 = xr.DataArray(
+    data=[[nan, -2.0, 1.0, 11.0]],
+    dims=["date", "station"],
+    coords=dict(date=["01"], station=[100, 101, 102, 103]),
+)
+
+DA_A = xr.DataArray(
+    data=[-1.0, 0.0, -1.0, 10.0],
+    dims=["station"],
+    coords=dict(station=[100, 101, 102, 103]),
+)
+
+DA_B = xr.DataArray(
+    data=[2.0, 2.0, 0.0, 12.0],
+    dims=["station"],
+    coords=dict(station=[100, 101, 102, 103]),
+)
+
+EXP_G_J_RECT2 = xr.DataArray(
+    data=[[nan, 0.0, 1.0, 1.0]],
+    dims=["date", "station"],
+    coords=dict(date=["01"], station=[100, 101, 102, 103]),
+)
+
+EXP_PHI_J_RECT1 = xr.DataArray([nan, 0.0, 8.0, 54.0])  # a = -1, b = 2
+
+EXP_PHI_J_RECT2 = xr.DataArray(
+    data=[[nan, 0.0, 6.0, 2.0]],
+    dims=["date", "station"],
+    coords=dict(date=["01"], station=[100, 101, 102, 103]),
+)
+
+EXP_PHI_J_PRIME_RECT1 = 4 * EXP_G_J_RECT1
+
+DA_X3 = xr.DataArray([nan, -3.0, 0.0, 2.0, 5.0, 10.0])
+# s = 5
+# a = -2, b = 1, c = 5, d = 8
+EXP_G_J_TRAP1 = xr.DataArray([nan, 0.0, 2 / 3, 2.5, 7 - 3 / 2, 7.0])
+
+DA_A_TRAP = xr.DataArray(
+    data=[[0, 10]],
+    dims=["date", "station"],
+    coords=dict(date=["01"], station=[100, 101]),
+)
+
+DA_B_TRAP = xr.DataArray(
+    data=[[2, 20]],
[…]
