+++ src/scores/continuous/__init__.py
@@ -22,6 +22,13 @@ from scores.continuous.standard_impl import (
     multiplicative_bias,
     rmse,
 )
+from scores.continuous.threshold_weighted_impl import (
+    tw_absolute_error,
+    tw_expectile_score,
+    tw_huber_loss,
+    tw_quantile_score,
+    tw_squared_error,
+)
 from scores.processing.isoreg_impl import isotonic_fit
 
 __all__ = [
@@ -41,4 +48,9 @@ __all__ = [
     "consistent_expectile_score",
     "consistent_huber_score",
     "consistent_quantile_score",
+    "tw_absolute_error",
+    "tw_expectile_score",
+    "tw_huber_loss",
+    "tw_quantile_score",
+    "tw_squared_error",
 ]
