### 2024-08-02 02eafeb
Update release notes to track upcoming version 1.1.0
