### 2024-08-02 4cca3a0
List the addition of threshold weighted scores tutorial in release notes
