### 2024-08-02 cbcd2db
Add basic details for each release from 0.0.1 to 0.9.2 to release notes
