### 2024-08-02 2b75fb7
Minor updates to maintainers notes (#617)
* Remove quotation mark

* Update numbering in maintainers notes
