### 2024-08-02 c3a538b
Change "trapezoidal" to "numpy.trapezoid"
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
