### 2024-08-02 c946d7d
Update release notes (#616)
* Update release notes 
---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
