### 2024-08-02 5d2c442
Relocates one paragraph in maintainers notes
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
