### 2024-08-02 b0f99b1
Improved docstring rendering for murphy_thetas and murphy_score
