### 2024-08-02 7585beb
Improve docstring rendering Add check_args documentation
