+++ src/scores/probability/brier_impl.py
@@ -25,7 +25,7 @@ def brier_score(
     Brier score, see https://en.wikipedia.org/wiki/Brier_score.
 
     If you want to speed up performance with Dask and are confident of your input data,
-    or if you want observations to take values between 0 and 1, set `check_inputs` to `False`.
+    or if you want observations to take values between 0 and 1, set ``check_args`` to ``False``.
 
     Args:
         fcst: Forecast or predicted variables in xarray.
@@ -41,9 +41,10 @@ def brier_score(
             forecast and observed dimensions must match precisely.
         weights: Optionally provide an array for weighted averaging (e.g. by area, by latitude,
             by population, custom)
+        check_args: will perform some tests on the data if set to True
     Raises:
-        ValueError: if `fcst` contains non-nan values outside of the range [0, 1]
-        ValueError: if `obs` contains non-nan values not in the set {0, 1}
+        ValueError: if ``fcst`` contains non-nan values outside of the range [0, 1]
+        ValueError: if ``obs`` contains non-nan values not in the set {0, 1}
     """
     if check_args:
         error_msg = ValueError("`fcst` contains values outside of the range [0, 1]")
