### 2024-08-02 790d4d4
Add mathjax to the brier score API docstring
