+++ docs/release_notes.md
@@ -2,32 +2,30 @@
 
 ## Version 1.1.0 (Upcoming Release) 
 
-For the full details of all changes in this release, see the [GitHub commit history](https://github.com/nci/scores/compare/1.0.0...develop). Below are the changes we think users may wish to be aware of.
+For a list of all changes in this release, see the [full changelog](https://github.com/nci/scores/compare/1.0.0...develop). Below are the changes we think users may wish to be aware of.
 
 ### Features
 
-- Added five new threshold weighted scores 
+- Added five new scores 
 	- threshold weighted squared error: ` scores.continuous.tw_squared_error`
 	- threshold weighted absolute error: `scores.continuous.tw_absolute_error`
 	- threshold weighted quantile score: `scores.continuous.tw_quantile_score`
 	- threshold weighted expectile score: `scores.continuous.tw_expectile_score`
 	- threshold weighted Huber loss: `scores.continuous.tw_huber_loss`.  
-See [PR #609](https://github.com/nci/scores/pull/609) by [@nicholasloveday](https://github.com/nicholasloveday).
-
-### Breaking Changes
-
-### Deprecations
-
-### Bug Fixes
+See [PR #609](https://github.com/nci/scores/pull/609).
 
 ### Documentation
 
-- Added "Threshold Weighted Scores" tutorial. See [PR #609](https://github.com/nci/scores/pull/609) by [@nicholasloveday](https://github.com/nicholasloveday).
-- Removed nbviewer link from documentation. See [PR #615](https://github.com/nci/scores/pull/615) by [@tennlee](https://github.com/tennlee).
+- Added "Threshold Weighted Scores" tutorial. See [PR #609](https://github.com/nci/scores/pull/609).
+- Removed nbviewer link from documentation. See [PR #615](https://github.com/nci/scores/pull/615).
 
 ### Internal Changes
 
-- Modified `numpy.trapezoid` call to work with either NumPy 1 or 2. See [PR #610](https://github.com/nci/scores/pull/610) by [@tennlee](https://github.com/tennlee).
+- Modified `numpy.trapezoid` call to work with either NumPy 1 or 2. See [PR #610](https://github.com/nci/scores/pull/610).
+
+### Contributors to this Release
+
+Nicholas Loveday ([@nicholasloveday](https://github.com/nicholasloveday)), Tennessee Leeuwenburg ([@tennlee](https://github.com/tennlee)), Stephanie Chong ([@Steph-Chong](https://github.com/Steph-Chong)) and Robert J. Taggart ([@rob-taggart](https://github.com/rob-taggart)). 
 
 ## Version 1.0.0 (July 10, 2024)
 
@@ -57,86 +55,90 @@ year = {2024}
 }
 ```
 
-For the full details of all changes in this release, see the [GitHub commit history](https://github.com/nci/scores/compare/0.9.3...1.0.0). 
+For a list of all changes in this release, see the [full changelog](https://github.com/nci/scores/compare/0.9.3...1.0.0). 
 
 ## Version 0.9.3 (July 9, 2024)
 
-For the full details of all changes in this release, see the [GitHub commit history](https://github.com/nci/scores/compare/0.9.2...0.9.3). Below are the changes we think users may wish to be aware of.
+For a list of all changes in this release, see the [full changelog](https://github.com/nci/scores/compare/0.9.2...0.9.3). Below are the changes we think users may wish to be aware of.
 
 ### Breaking Changes
 
-- Renamed and relocated function `scores.continuous.correlation` to `scores.continuous.correlation.pearsonr`. See [PR #583](https://github.com/nci/scores/pull/583) by [@nicholasloveday](https://github.com/nicholasloveday). 
[…]
