+++ docs/release_notes.md
@@ -14,6 +14,12 @@ For a list of all changes in this release, see the [full changelog](https://gith
 	- threshold weighted Huber loss: `scores.continuous.tw_huber_loss`.  
 See [PR #609](https://github.com/nci/scores/pull/609).
 
+### Breaking Changes
+
+### Deprecations
+
+### Bug Fixes
+
 ### Documentation
 
 - Added "Threshold Weighted Scores" tutorial. See [PR #609](https://github.com/nci/scores/pull/609).
