### 2024-08-05 7883fe9
Update formatting for second paragraph of README.md
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
