### 2024-08-05 64304bf
Minor update to readme (#625)
* Add journal article link early in README


Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
