### 2024-08-05 1c24a05
Update to release notes (#622)
* Update Release Notes
