### 2024-08-07 b916bd7
Add PyPI badge to README.md
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
