### 2024-08-07 6981d95
Add Conda-Forge badge
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
