### 2024-08-07 c0331dd
Add .svg to PyPI and conda-forge badge
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
