### 2024-08-07 c8d0ff6
Adds PyPI and conda-forge badges to README.md (#626)
* Add PyPI badge to README.md
* Add Conda-Forge badge

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
