### 2024-08-07 837abc5
Adjust capitalisation for consistency
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
