### 2024-08-07 e72e1e8
Updates to the documentation tech stack have resulted in errors and the docs could not be built. This change introduces a workaround, however there are still warnings being raised. It is likely that they will persist until the underlying libraries adapt to the new versions. This commit should restore the ability for readthedocs to build the documentation (hopefully).
