### 2024-08-07 39f5fc7
Add conda installation instructions to installation.md
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
