### 2024-08-07 5fd22b5
Add conda installation instructions to installation.md
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
