### 2024-08-08 5db29c9
Further updates to installation.md
