### 2024-08-09 94c9868
Adds the word "dependencies", in two spots, to aid clarity
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
