### 2024-08-09 57d89c4
Add "default" to installation section of README.md to aid clarity
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
