### 2024-08-09 c252593
Relocates some text in installation section of README.md
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
