### 2024-08-09 701657f
Add bold to one sentence in README.md
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
