### 2024-08-09 18eaf90
Add conda install instructions to installation.md and README.md, plus various updates to README.md (#627)
* Add conda installation instructions to installation.md among other updates
* Change path to environment to named environment
* Updated installation instructions in README.md among other updates

---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
Co-authored-by: Nicholas Loveday <48701367+nicholasloveday@users.noreply.github.com>
