### 2024-08-09 6b0caa9
Add a missing ":" in the installation doc
