### 2024-08-09 ecb6af0
Update release notes for upcoming release
