### 2024-08-09 e6c0db2
Finalise release notes for release 1.1.0
