### 2024-08-09 dd97750
Update release_notes.md
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
