+++ src/scores/continuous/__init__.py
@@ -16,6 +16,7 @@ from scores.continuous.murphy_impl import murphy_score, murphy_thetas
 from scores.continuous.quantile_loss_impl import quantile_score
 from scores.continuous.standard_impl import (
     additive_bias,
+    additive_bias_percentage,
     mae,
     mean_error,
     mse,
@@ -44,6 +45,7 @@ __all__ = [
     "additive_bias",
     "mean_error",
     "multiplicative_bias",
+    "additive_bias_percentage",
     "isotonic_fit",
     "consistent_expectile_score",
     "consistent_huber_score",
