### 2024-08-21 abe4542
Minor formatting for compatible with black (#613-pbias)
