### 2024-08-21 dbb6dfc
Further Minor formatting for compatible with black (#613-pbias)
