### 2024-08-22 810bbc3
Adding scores to api and included list of metrics (#613-pbias)
