### 2024-08-23 3ea57f1
add test raises
