+++ src/scores/probability/crps_impl.py
@@ -920,6 +920,7 @@ def tail_twcrps_for_ensemble(
     twCRPS but for an upper or lower tail above or below the `threshold` arg.
     """
     if tail == "upper":
+
         def _vfunc(x):
             return np.maximum(x, threshold)
 
