+++ tests/probabilty/test_crps.py
@@ -18,8 +18,8 @@ from scores.probability import (
     crps_cdf,
     crps_cdf_brier_decomposition,
     crps_for_ensemble,
-    twcrps_for_ensemble,
     tail_twcrps_for_ensemble,
+    twcrps_for_ensemble,
 )
 from scores.probability.crps_impl import (
     crps_cdf_exact,
