+++ src/scores/probability/__init__.py
@@ -10,8 +10,8 @@ from scores.probability.crps_impl import (
     crps_cdf_brier_decomposition,
     crps_for_ensemble,
     crps_step_threshold_weight,
-    twcrps_for_ensemble,
     tail_twcrps_for_ensemble,
+    twcrps_for_ensemble,
 )
 from scores.probability.roc_impl import roc_curve_data
 from scores.processing.isoreg_impl import isotonic_fit
