### 2024-08-24 ec8d555
With the introduction of numpy 2.1, some dask operations now compute data types to a numpy generic (e.g. a scalar, such as int or float) rather than to an array of such things. This change modifies the test to incorporate a union type of array and generic in assert statements for dask operations
