### 2024-08-25 fb2a49a
Modify type test for compatibility with Python 3.9
