+++ tests/probabilty/crps_test_data.py
@@ -573,6 +573,11 @@ DA_FCST_CRPSENS = xr.DataArray(
 DA_OBS_CRPSENS = xr.DataArray(data=[2.0, 3, 1, np.nan, 4], dims=["stn"], coords={"stn": [101, 102, 103, 104, 105]})
 DA_WT_CRPSENS = xr.DataArray(data=[1, 2, 1, 0, 2], dims=["stn"], coords={"stn": [101, 102, 103, 104, 105]})
 DA_T_TWCRPSENS = xr.DataArray(data=[np.nan, 1, 10, 1, -2], dims=["stn"], coords={"stn": [101, 102, 103, 104, 105]})
+
+DS_FCST_CRPSENS = xr.Dataset({"a": DA_FCST_CRPSENS, "b": DA_FCST_CRPSENS})
+DS_OBS_CRPSENS = xr.Dataset({"a": DA_OBS_CRPSENS, "b": DA_OBS_CRPSENS})
+DS_WT_CRPSENS = xr.Dataset({"a": DA_WT_CRPSENS, "b": DA_WT_CRPSENS})
+DS_T_TWCRPSENS = xr.Dataset({"a": DA_T_TWCRPSENS, "b": DA_T_TWCRPSENS})
 # first and second (spread) terms from crps for ensembles, "ecdf" and "fair" methods
 FIRST_TERM = xr.DataArray(
     data=[10 / 4, 8 / 4, 4 / 3, np.nan, 2], dims=["stn"], coords={"stn": [101, 102, 103, 104, 105]}
@@ -593,48 +598,52 @@ EXP_CRPSENS_ECDF = FIRST_TERM - SPREAD_ECDF
 EXP_CRPSENS_FAIR = FIRST_TERM - SPREAD_FAIR
 EXP_CRPSENS_WT = (EXP_CRPSENS_ECDF * DA_WT_CRPSENS).mean("stn")
 
+EXP_CRPSENS_ECDF_DS = xr.Dataset({"a": EXP_CRPSENS_ECDF, "b": EXP_CRPSENS_ECDF})
+EXP_CRPSENS_FAIR_DS = xr.Dataset({"a": EXP_CRPSENS_FAIR, "b": EXP_CRPSENS_FAIR})
+EXP_CRPSENS_WT_DS = xr.Dataset({"a": EXP_CRPSENS_WT, "b": EXP_CRPSENS_WT})
+
 # exp test data for twCRPS with upper tail for thresholds >=1
-UPPER_TAIL_FIRST_TERM = xr.DataArray(
+UPPER_TAIL_FIRST_TERM_DA = xr.DataArray(
     data=[9 / 4, 6 / 4, 3 / 3, np.nan, 2], dims=["stn"], coords={"stn": [101, 102, 103, 104, 105]}
 )
-UPPER_TAIL_SPREAD_ECDF = xr.DataArray(
+UPPER_TAIL_SPREAD_ECDF_DA = xr.DataArray(
     data=[(11 + 7 + 7 + 13) / 32, (4 + 4 + 4 + 8) / 32, (3 + 3 + 6) / 18, np.nan, 0],
     dims=["stn"],
     coords={"stn": [101, 102, 103, 104, 105]},
 )
-UPPER_TAIL_SPREAD_FAIR = xr.DataArray(
+UPPER_TAIL_SPREAD_FAIR_DA = xr.DataArray(
     data=[(11 + 7 + 7 + 13) / 24, (4 + 4 + 4 + 8) / 24, (3 + 3 + 6) / 12, np.nan, np.nan],
     dims=["stn"],
[…]
