+++ tests/probabilty/test_crps.py
@@ -618,24 +618,22 @@ def test_crps_for_ensemble():
 
 def test_crps_for_ensemble_ds():
     """Tests `crps_for_ensemble` returns as expected with a DataSet for the input"""
-    fcst = xr.Dataset({"a": crps_test_data.DA_FCST_CRPSENS, "b": crps_test_data.DA_FCST_CRPSENS})
-    obs = xr.Dataset({"a": crps_test_data.DA_OBS_CRPSENS, "b": crps_test_data.DA_OBS_CRPSENS})
-    exp_ecdf = xr.Dataset({"a": crps_test_data.EXP_CRPSENS_ECDF, "b": crps_test_data.EXP_CRPSENS_ECDF})
-    exp_fair = xr.Dataset({"a": crps_test_data.EXP_CRPSENS_FAIR, "b": crps_test_data.EXP_CRPSENS_FAIR})
-    exp_weighted_mean = xr.Dataset({"a": crps_test_data.EXP_CRPSENS_WT, "b": crps_test_data.EXP_CRPSENS_WT})
-
-    result_ecdf = crps_for_ensemble(fcst, obs, "ens_member", method="ecdf", preserve_dims="all")
-    result_fair = crps_for_ensemble(fcst, obs, "ens_member", method="fair", preserve_dims="all")
+    result_ecdf = crps_for_ensemble(
+        crps_test_data.DS_FCST_CRPSENS, crps_test_data.DS_OBS_CRPSENS, "ens_member", method="ecdf", preserve_dims="all"
+    )
+    result_fair = crps_for_ensemble(
+        crps_test_data.DS_FCST_CRPSENS, crps_test_data.DS_OBS_CRPSENS, "ens_member", method="fair", preserve_dims="all"
+    )
     result_weighted_mean = crps_for_ensemble(
-        fcst,
-        obs,
+        crps_test_data.DS_FCST_CRPSENS,
+        crps_test_data.DS_OBS_CRPSENS,
         "ens_member",
         method="ecdf",
-        weights=crps_test_data.DA_WT_CRPSENS,
+        weights=crps_test_data.DS_WT_CRPSENS,
     )
-    xr.testing.assert_allclose(result_ecdf, exp_ecdf)
-    xr.testing.assert_allclose(result_fair, exp_fair)
-    xr.testing.assert_allclose(result_weighted_mean, exp_weighted_mean)
+    xr.testing.assert_allclose(result_ecdf, crps_test_data.EXP_CRPSENS_ECDF_DS)
+    xr.testing.assert_allclose(result_fair, crps_test_data.EXP_CRPSENS_FAIR_DS)
+    xr.testing.assert_allclose(result_weighted_mean, crps_test_data.EXP_CRPSENS_WT_DS)
 
 
 def test_crps_for_ensemble_raises():
@@ -665,25 +663,119 @@ def test_crps_for_ensemble_dask():
 
 
 @pytest.mark.parametrize(
-    ("method", "tail", "threshold", "preserve_dims", "reduce_dims", "weights", "expected"),
+    ("fcst", "obs", "method", "tail", "threshold", "preserve_dims", "reduce_dims", "weights", "expected"),
     [
-        ("ecdf", "upper", 1, "all", None, None, crps_test_data.EXP_UPPER_TAIL_CRPSENS_ECDF),
-        ("fair", "upper", 1, "all", None, None, crps_test_data.EXP_UPPER_TAIL_CRPSENS_FAIR),
-        ("ecdf", "lower", 1, "all", None, None, crps_test_data.EXP_LOWER_TAIL_CRPSENS_ECDF),
-        ("fair", "lower", 1, "all", None, None, crps_test_data.EXP_LOWER_TAIL_CRPSENS_FAIR),
+        (
[…]

[…]
