+++ tests/probabilty/crps_test_data.py
@@ -618,6 +618,7 @@ UPPER_TAIL_SPREAD_FAIR_DA = xr.DataArray(
 )
 EXP_UPPER_TAIL_CRPSENS_ECDF_DA = UPPER_TAIL_FIRST_TERM_DA - UPPER_TAIL_SPREAD_ECDF_DA
 EXP_UPPER_TAIL_CRPSENS_FAIR_DA = UPPER_TAIL_FIRST_TERM_DA - UPPER_TAIL_SPREAD_FAIR_DA
+EXP_UPPER_TAIL_CRPSENS_ECDF_DS = xr.Dataset({"a": EXP_UPPER_TAIL_CRPSENS_ECDF_DA, "b": EXP_UPPER_TAIL_CRPSENS_ECDF_DA})
 
 # exp test data for twCRPS with lower tail for thresholds <=1
 LOWER_TAIL_FIRST_TERM_DA = xr.DataArray(
