### 2024-08-26 ed12560
first pass at twcrps docstring
