+++ src/scores/probability/crps_impl.py
@@ -888,9 +888,74 @@ def twcrps_for_ensemble(
     weights: Optional[XarrayLike] = None,
 ) -> xr.DataArray:
     """
-    Calculates the threshold weighted continuous ranked probability score (twCRPS) given ensemble
-    input using a chaing function ``vfunc``. An ensemble of forecasts can also be thought of as a random sample from the predictive
-    distribution.
+    Calculates the threshold weighted continuous ranked probability score (twCRPS) given
+    ensemble input using a chaing function ``vfunc``. An ensemble of forecasts can also
+    be thought of as a random sample from the predictive distribution.
+
+    The twCRPS is calculated by the formula
+    .. math::
+        \\text{CRPS}(F, y) = \\mathbb{E}_F \\left| v(X) - v(y) \\right|
+        - \\frac{1}{2} \\mathbb{E}_F \\left| v(X) - v(X') \\right|
+
+    where :math:`X` and :math:`X'` are independent samples of the predictive distribution :math:`F`,
+    :math:`y` is the observation, and :math:`v` is a 'chaining function'.
+
+    The chaining function :math:`v` is the antiderivative of the weight function :math:`w`: which is a
+    non-negative weight function that assigns a weight to each threshold value. For example, if we
+    wanted to give a weight of 1 to thresholds above threshold :math:`t` and a weight of 0 to
+    thresholds below :math:`t`, our weight function would be :math:`w(x) = \\mathbb{1}{x > t}`,
+    where :math:`\\mathbb{1}` is the indicator function which returns a value of 1 if the condition
+    is true and 0 otherwise. The chaining function would then be :math:`v(x) = \\text{max}(x, t)`.
+
+    The ensemble representation of the emperical twCRPS is
+    .. math::
+        \\text{twCRPS}(F_{\\text{ens}}, y; v) = \\frac{1}{M} \\sum_{m=1}^{M} \\left| v(x_m) - v(y) \\right|
+        - \\frac{1}{2M^2} \\sum_{m=1}^{M} \\sum_{j=1}^{M} \\left| v(x_m) - v(x_j) \\right|,
+
+    where :math:`M` is the number of ensemble members and :math:`x_m` is the :math:`m`-th ensemble member.
+
+    Args:
+        fcst: Forecast data. Must have a dimension `ensemble_member_dim`.
+        obs: Observation data.
+        ensemble_member_dim: the dimension that specifies the ensemble member or the sample
+            from the predictive distribution.
+        v_func: the chaining function.
+        method: Either "ecdf" or "fair".
+        reduce_dims: Dimensions to reduce. Can be "all" to reduce all dimensions.
+        preserve_dims: Dimensions to preserve. Can be "all" to preserve all dimensions.
+        weights: Weights for calculating a weighted mean of individual scores.
+
+    Returns:
+        xarray object of twCRPS values.
+
+    Raises:
+        ValueError: when method is not one of "ecdf" or "fair".
+
+    Notes:
+        Chaining functions can be created to vary the weights across given dimensions
+        such as varying the weights by climatological values.
+
+    References:
+        Allen, S., Ginsbourger, D., & Ziegel, J. (2023). Evaluating forecasts for high-impact
+        events using transformed kernel scores. SIAM/ASA Journal on Uncertainty
+        Quantification, 11(3), 906-940. https://doi.org/10.1137/22M1532184
+
+    See also:
+        :py:func:`scores.probability.crps_cdf`
+        :py:func:`tail_twcrps_for_ensemble`
+
+    Examples:
+        Calculate the twCRPS for an ensemble of forecasts where the chaining function is the
[…]
