+++ src/scores/probability/crps_impl.py
@@ -826,6 +826,8 @@ def crps_for_ensemble(
 
     See also:
         :py:func:`scores.probability.crps_cdf`
+        :py:func:`twcrps_for_ensemble`
+        :py:func:`tail_twcrps_for_ensemble`
 
     References:
         - C. Ferro (2014), "Fair scores for ensemble forecasts", Quarterly Journal of the \
@@ -880,7 +882,7 @@ def twcrps_for_ensemble(
     fcst: XarrayLike,
     obs: XarrayLike,
     ensemble_member_dim: str,
-    v_func: Callable[[Union[XarrayLike, float]], XarrayLike],
+    v_func: Callable[[XarrayLike], XarrayLike],
     *,  # Force keywords arguments to be keyword-only
     method: Literal["ecdf", "fair"] = "ecdf",
     reduce_dims: Optional[Sequence[str]] = None,
@@ -893,26 +895,29 @@ def twcrps_for_ensemble(
     be thought of as a random sample from the predictive distribution.
 
     The twCRPS is calculated by the formula
+
+
     .. math::
-        \\text{CRPS}(F, y) = \\mathbb{E}_F \\left| v(X) - v(y) \\right|
-        - \\frac{1}{2} \\mathbb{E}_F \\left| v(X) - v(X') \\right|,
+        \\text{CRPS}(F, y) = \\mathbb{E}_F \\left| v(X) - v(y) \\right| - \\frac{1}{2} \\mathbb{E}_F \\left| v(X) - v(X') \\right|,
 
     where :math:`X` and :math:`X'` are independent samples of the predictive distribution :math:`F`,
     :math:`y` is the observation, and :math:`v` is a 'chaining function'.
 
-    The chaining function :math:`v` is the antiderivative of the weight function :math:`w`: which is a
-    non-negative weight function that assigns a weight to each threshold value. For example, if we
-    wanted to give a weight of 1 to thresholds above threshold :math:`t` and a weight of 0 to
-    thresholds below :math:`t`, our weight function would be :math:`w(x) = \\mathbb{1}{x > t}`,
+    The chaining function :math:`v` is the antiderivative of the weight function :math:`w`, which is a
+    continous, non-negative function that assigns a weight to each threshold value. For example, if we
+    wanted to assign a weight of 1 to thresholds above threshold :math:`t` and a weight of 0 to
+    thresholds below :math:`t`, our weight function would be :math:`w(x) = \\mathbb{1}{(x > t)}`,
     where :math:`\\mathbb{1}` is the indicator function which returns a value of 1 if the condition
     is true and 0 otherwise. The chaining function would then be :math:`v(x) = \\text{max}(x, t)`.
 
     The ensemble representation of the emperical twCRPS is
-    .. math::
-        \\text{twCRPS}(F_{\\text{ens}}, y; v) = \\frac{1}{M} \\sum_{m=1}^{M} \\left| v(x_m) - v(y) \\right|
-        - \\frac{1}{2M^2} \\sum_{m=1}^{M} \\sum_{j=1}^{M} \\left| v(x_m) - v(x_j) \\right|,
 
-    where :math:`M` is the number of ensemble members and :math:`x_m` is the :math:`m`-th ensemble member.
+
+    .. math::
+        \\text{twCRPS}(F_{\\text{ens}}, y; v) = \\frac{1}{M} \\sum_{m=1}^{M} \\left| v(x_m) - v(y) \\right| -
+        \\frac{1}{2M^2} \\sum_{m=1}^{M} \\sum_{j=1}^{M} \\left| v(x_m) - v(x_j) \\right|,
+
+    where :math:`M` is the number of ensemble members.
 
     Args:
[…]
