### 2024-08-26 3b051e7
add broadcasting tests for twcrps
