+++ tests/probabilty/crps_test_data.py
@@ -570,6 +570,20 @@ DA_FCST_CRPSENS = xr.DataArray(
     dims=["stn", "ens_member"],
     coords={"stn": [101, 102, 103, 104, 105], "ens_member": [1, 2, 3, 4]},
 )
+DA_FCST_CRPSENS_LT = xr.DataArray(
+    data=[
+        [[0.0, 4, 3, 7], [1, -1, 2, 4], [0, 1, 4, np.nan], [2, 3, 4, 1], [2, np.nan, np.nan, np.nan]],
+        [
+            [2, 2, 2, 2],
+            [np.nan, np.nan, np.nan, np.nan],
+            [np.nan, np.nan, np.nan, np.nan],
+            [np.nan, np.nan, np.nan, np.nan],
+            [np.nan, np.nan, np.nan, np.nan],
+        ],
+    ],
+    dims=["lead_time", "stn", "ens_member"],
+    coords={"stn": [101, 102, 103, 104, 105], "ens_member": [1, 2, 3, 4], "lead_time": [1, 2]},
+)
 DA_OBS_CRPSENS = xr.DataArray(
     data=[2.0, 3, 1, np.nan, 4, 5], dims=["stn"], coords={"stn": [101, 102, 103, 104, 105, 106]}
 )
@@ -582,7 +596,9 @@ DS_WT_CRPSENS = xr.Dataset({"a": DA_WT_CRPSENS, "b": DA_WT_CRPSENS})
 DS_T_TWCRPSENS = xr.Dataset({"a": DA_T_TWCRPSENS, "b": DA_T_TWCRPSENS})
 # first and second (spread) terms from crps for ensembles, "ecdf" and "fair" methods
 FIRST_TERM = xr.DataArray(
-    data=[10 / 4, 8 / 4, 4 / 3, np.nan, 2], dims=["stn"], coords={"stn": [101, 102, 103, 104, 105]}
+    data=[10 / 4, 8 / 4, 4 / 3, np.nan, 2],
+    dims=["stn"],
+    coords={"stn": [101, 102, 103, 104, 105]},
 )
 SPREAD_ECDF = xr.DataArray(
     data=[(14 + 8 + 8 + 14) / 32, (6 + 10 + 6 + 10) / 32, (5 + 4 + 7) / 18, np.nan, 0],
@@ -604,6 +620,22 @@ EXP_CRPSENS_ECDF_DS = xr.Dataset({"a": EXP_CRPSENS_ECDF, "b": EXP_CRPSENS_ECDF})
 EXP_CRPSENS_FAIR_DS = xr.Dataset({"a": EXP_CRPSENS_FAIR, "b": EXP_CRPSENS_FAIR})
 EXP_CRPSENS_WT_DS = xr.Dataset({"a": EXP_CRPSENS_WT, "b": EXP_CRPSENS_WT})
 
+# Test data for broadcasting
+FIRST_TERM_BC = xr.DataArray(
+    data=[[10 / 4, 8 / 4, 4 / 3, np.nan, 2], [0, np.nan, np.nan, np.nan, np.nan]],
+    dims=["lead_time", "stn"],
+    coords={"lead_time": [1, 2], "stn": [101, 102, 103, 104, 105]},
+)
+SPREAD_ECDF_BC = xr.DataArray(
+    data=[
+        [(14 + 8 + 8 + 14) / 32, (6 + 10 + 6 + 10) / 32, (5 + 4 + 7) / 18, np.nan, 0],
+        [0, np.nan, np.nan, np.nan, np.nan],
+    ],
+    dims=["lead_time", "stn"],
[…]
