### 2024-08-30 8826e2d
Bump jupyterlab from 4.1.5 to 4.2.5 in /.binder
Bumps [jupyterlab](https://github.com/jupyterlab/jupyterlab) from 4.1.5 to 4.2.5.
- [Release notes](https://github.com/jupyterlab/jupyterlab/releases)
- [Changelog](https://github.com/jupyterlab/jupyterlab/blob/@jupyterlab/lsp@4.2.5/CHANGELOG.md)
- [Commits](https://github.com/jupyterlab/jupyterlab/compare/@jupyterlab/lsp@4.1.5...@jupyterlab/lsp@4.2.5)

---
updated-dependencies:
- dependency-name: jupyterlab
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
