+++ docs/included.md
@@ -13,10 +13,6 @@
   - [API](api.md#scores.continuous.additive_bias)
   - [Tutorial](project:./tutorials/Additive_and_multiplicative_bias.md)
   - [Mean Error (WWRP/WGNE Joint Working Group on Forecast Verification Research)](https://www.cawcr.gov.au/projects/verification/#meanerror)
-* - Additive Bias percentage
-  - [API](api.md#scores.continuous.additive_bias_percentage)
-  - [Tutorial](project:./tutorials/Additive_and_multiplicative_bias.md)
-  - [https://search.r-project.org/CRAN/refmans/hydroGOF/html/pbias.html)  
 * - Consistent Expectile Score
   - [API](api.md#scores.continuous.consistent_expectile_score)
   - [Tutorial](project:./tutorials/Consistent_Scores.md)
@@ -101,6 +97,10 @@
   - [API](api.md#scores.continuous.correlation.pearsonr)
   - [Tutorial](project:./tutorials/Pearsons_Correlation.md)
   - [Wikipedia](https://en.wikipedia.org/wiki/Pearson_correlation_coefficient)
+* - Percent Bias
+  - [API](api.md#scores.continuous.pbias)
+  - [Tutorial](project:./tutorials/Additive_and_multiplicative_bias.md)
+  - [Percent Bias (CRAN hydroGOF)](https://search.r-project.org/CRAN/refmans/hydroGOF/html/pbias.html); [Sorooshian et al. (1993)](https://doi.org/10.1029/92WR02617); [Alfieri et al. (2014)](http://dx.doi.org/10.1016/j.jhydrol.2014.06.035); [Dawson et al. (2007)](https://doi.org/10.1016/j.envsoft.2006.06.008); [Moriasi et al. (2007)[https://doi.org/10.13031/2013.23153]
 * - Pinball Loss, *see Quantile Loss*
   - &mdash;
   - &mdash;
