+++ tests/probabilty/test_crps.py
@@ -992,7 +992,7 @@ def test_twcrps_for_ensemble(fcst, obs, method, v_func, preserve_dims, reduce_di
         fcst,
         obs,
         ensemble_member_dim="ens_member",
-        v_func=v_func,
+        chaining_func=v_func,
         method=method,
         preserve_dims=preserve_dims,
         reduce_dims=reduce_dims,
@@ -1007,7 +1007,7 @@ def test_twcrps_for_ensemble_dask():
         fcst=crps_test_data.DA_FCST_CRPSENS.chunk(),
         obs=crps_test_data.DA_OBS_CRPSENS.chunk(),
         ensemble_member_dim="ens_member",
-        v_func=v_func1,
+        chaining_func=v_func1,
         method="ecdf",
         preserve_dims="all",
         reduce_dims=None,
@@ -1023,7 +1023,7 @@ def test_twcrps_for_ensemble_dask():
         fcst=crps_test_data.DS_FCST_CRPSENS.chunk(),
         obs=crps_test_data.DS_OBS_CRPSENS.chunk(),
         ensemble_member_dim="ens_member",
-        v_func=v_func1,
+        chaining_func=v_func1,
         method="ecdf",
         preserve_dims="all",
         reduce_dims=None,
