### 2024-09-04 cb9d0b3
Allow use of slightly earlier pandas versions
