+++ pyproject.toml
@@ -21,7 +21,7 @@ classifiers = [
 ]
 dependencies = [
     "xarray ~= 2024.1",
-    "pandas ~= 2.2",
+    "pandas ~= 2.0",
     "scipy ~= 1.1",
     "bottleneck ~= 1.3",
     "scikit-learn ~= 1.4",
