### 2024-09-04 7966a97
Allow use of slightly earlier pandas versions (#652)
* Allow use of slightly earlier pandas versions

* Merge pull request #643 from tennlee/dask_type_checking

Dask type checking
