### 2024-09-05 3f5d540
fix conda based install contributing instructions
