+++ docs/contributing.md
@@ -91,7 +91,7 @@ pytest
 ```bash
 # overwrite default name `scoresenv` with `-n <new-name>` if desired
 conda env create -f environment.yml
-conda activate scoresenv
+conda activate scores
 pip install -e .[all]
 pytest
 ```
