### 2024-09-05 b92bb2a
Update docs/contributing.md
Co-authored-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
Signed-off-by: Maree Carroll <maree.carroll@gmail.com>
