### 2024-09-05 a03ee9d
Minor updates to pbias API docstrings
