+++ docs/included.md
@@ -221,11 +221,11 @@
   - &mdash;
   - &mdash;
 * - Tail Threshold Weighted Continuous Ranked Probability Score (twCRPS) for Ensembles
-  - [API](api.md#scores.probability.tail_twcrps_for_ensemble)   
+  - [API](api.md#scores.probability.tail_tw_crps_for_ensemble)   
   - &mdash;
   - [Allen et al. (2023)](https://doi.org/10.1137/22M1532184) 
 * - Threshold Weighted Continuous Ranked Probability Score (twCRPS) for Ensembles
-  - [API](api.md#scores.probability.twcrps_for_ensemble)   
+  - [API](api.md#scores.probability.tw_crps_for_ensemble)   
   - &mdash;
   - [Allen et al. (2023)](https://doi.org/10.1137/22M1532184) 
 ```
