### 2024-09-05 f5d90e1
improvements to docstrings
