### 2024-09-06 dd77f30
Update comment in environment.yml
