### 2024-09-06 63e992f
Update count of number of included metrics
