### 2024-09-06 365b818
Update count of number of included metrics (#659)
