### 2024-09-06 78fa173
Add backticks around np.inf
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
