### 2024-09-06 5086e60
Remove ".dx" from one DOI
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
