### 2024-09-06 b21cf4f
Adjust "*" to \\cdot in mathjax
