### 2024-09-06 43b2bd9
Minor updates to pbias API docstring (#655)
* Updates to pbias API docstrings
---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
Co-authored-by: Tennessee Leeuwenburg <tennessee.leeuwenburg@bom.gov.au>
