### 2024-09-07 170c503
Updates to use https links in references and other doc adjustments
