### 2024-09-08 #6889 @Steph-Chong
I was just wondering if there has been any update on the Crossref metadata? The Crossref metadata is still showing the newlines.
