### 2024-09-09 2480bc1
firm can now take a list of xr.datarrays for the thresholds (#661)
