### 2024-09-09 11692bc
add chaining_func_kwargs
