### 2024-09-10 e7a52c9
Update continuous and categorical sections of table in README
