### 2024-09-10 1d32742
Further update to full list sentence
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
