### 2024-09-10 f06795c
Remove extra bracket
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
