### 2024-09-10 fafc2ad
Swapped order of two sentences
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
