### 2024-09-10 f402b65
Swapping the order of sentences back again
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
