### 2024-09-11 41e0541
Adds link to "Dimension Handling" notebook
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
