### 2024-09-11 3bd0750
Update Categorical section of table to resolve conflicts
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
