### 2024-09-11 33bcd86
Update continuous and categorical sections of table in README (#665)
* Update continuous and categorical sections of table in README
* Adding link to full list of metrics above table

---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
