### 2024-09-11 ad55878
Update Key Benefits page (#666)
* Update Key Benefits page

* Adds link to "Dimension Handling" notebook

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>

---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
