### 2024-09-11 6ad938e
Update README.md
Adjust link text slightly

Signed-off-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
