### 2024-09-11 6603ddd
Update README.md
Tweak to acknowledgment heading

Signed-off-by: Tennessee Leeuwenburg <134973832+tennlee@users.noreply.github.com>
