### 2024-09-11 66b82d8
Remove one instance of "metric" and change one uppercase "S" to lowercase "s"
