### 2024-09-11 3e04878
Improve clarity of dot point about updating conda-based virtual environment installation instructions
