### 2024-09-11 7b5e5f4
Add information about asking new contributors to add their details to .zenodo.json
