+++ docs/maintainer.md
@@ -58,6 +58,24 @@ For the full details of all changes in this release, see the [GitHub commit hist
 For each entry: "Brief description. See [PR #ABC](add link) by [@username](add link)."
 ```
 
+## This section covers asking new contributors to add their details to .zenodo.json
+
+```
+Thank you very much for your contribution. 
+
+When we release a new version of `scores`, that version is archived on Zenodo. See: https://doi.org/10.5281/zenodo.12697241
+
+As you have contributed to `scores`, would you like to be listed on Zenodo as an author the next time `scores` is archived? 
+
+If so, please open a new pull request. In that pull request please add your details to .zenodo.json (which can be found in the  `scores` root directory).
+
+In .zenodo.json, please add your details at the bottom of the “creators” section. The fields you will need to complete are:
+
+1. “orcid”. This is an optional field. If you don’t have an ORCID, but would like one, you can obtain one here: https://info.orcid.org/researchers/ .
+2. “affiliation”. Options include: the name of the institution you are affiliated with, “Independent Researcher” or “Independent Contributor”.
+3. “name”. Format: surname, given name(s).
+```
+
 ## This section covers how to build the documentation locally 
 (Readthedocs should update automatically from a GitHub Action)
 
