### 2024-09-11 c8e5eb4
Update how to format release notes section of Maintainers Notes
