### 2024-09-11 2aee079
Add information about what to change in release notes
