### 2024-09-11 fb4d992
Add details about what needs to be checked in Zenodo (#674)
