### 2024-09-11 7620c11
Add comments on version compatbility to the maintainers guide
