### 2024-09-11 1fd3778
Add comments on version compatbility to the maintainers guide
