### 2024-09-12 df0324f
Update FIRM dotpoint for improved clarity
Co-authored-by: Nicholas Loveday <48701367+nicholasloveday@users.noreply.github.com>
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
