### 2024-09-12 faac41a
Update release notes (#669)
Co-authored-by: Nicholas Loveday <48701367+nicholasloveday@users.noreply.github.com>
Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>

---------

Signed-off-by: Stephanie Chong <168800785+Steph-Chong@users.noreply.github.com>
Co-authored-by: Nicholas Loveday <48701367+nicholasloveday@users.noreply.github.com>
