### 2024-09-12 8114562
KGE scores (#678-kge)
