+++ src/scores/continuous/standard_impl.py
@@ -2,8 +2,9 @@
 This module contains standard methods which may be used for continuous scoring
 """
 
-from typing import Optional
+from typing import Optional, Tuple, Union
 
+import numpy as np
 import xarray as xr
 
 import scores.functions
@@ -418,3 +419,137 @@ def pbias(
 
     _pbias = 100 * error.mean(dim=reduce_dims) / obs.mean(dim=reduce_dims)
     return _pbias
+
+
+def kge(
+    fcst: xr.DataArray,
+    obs: xr.DataArray,
+    *,
+    reduce_dims: Optional[FlexibleDimensionTypes] = None,
+    preserve_dims: Optional[FlexibleDimensionTypes] = None,
+    scaling_factors: Optional[Union[list[float], np.ndarray]] = None,
+    return_components: bool = False,
+) -> XarrayLike:
+    # pylint: disable=too-many-locals
+    """
+    Calculate the Kling-Gupta Efficiency (KGE) between observed and simulated (or forecast) values.
+
+    KGE is a performance metric that decomposes the error into three components:
+    correlation, bias, and variability.
+    It is computed as:
+
+    .. math::
+        \\text{KGE} = 1 - \\sqrt{\\left[s_\rho \\cdot (\\rho - 1)\\right]^2 + \\left[s_\\alpha \\cdot (\\alpha - 1)\\right]^2 + \\left[s_\\beta \\cdot (\\beta - 1)\\right]^2}
+        \\rho = \\frac{\\text{cov}(x, y)}{\\sigma_x \\sigma_y}
+        \\alpha = \\frac{\\sigma_x}{\\sigma_y}
+        \\beta = \\frac{\\mu_x}{\\mu_y}
+
+    where:
+        - :math:`\\rho`  = Pearson's correlation coefficient between observed and forecast values
+        - :math:`\\alpha` is the ratio of the standard deviations (variability ratio)
+        - :math:`\\beta` is the ratio of the means (bias)
+        - :math:`x` and :math:'y' are forecast and observed values, respectively
+        - :math:`\\mu_x` and :math:`\\mu_y` are the means of forecast and observed values, respectively # pylint: disable=line-too-long
+        - :math:`\\sigma_x` and :math:`\\sigma_y` are the standard deviations of forecast and observed values, respectively # pylint: disable=line-too-long
+        - :math:`s_\\rho`, :math:'s_\\alpha and :math:`s_\\beta` are the scaling factors for
+            for the correlation coefficient :math:`\\rho`
+            for the variability term :math:`\\alpha`
+            for the bias term :math:`\\beta`
+
+    Args:
+        fcst: Forecast or predicted variables.
+        obs: Observed variables.
+        reduce_dims: Optionally specify which dimensions to reduce when
+            calculating the KGE. All other dimensions will be preserved.
+        preserve_dims: Optionally specify which dimensions to preserve when
+            calculating the KGE. All other dimensions will be reduced. As a
+            special case, 'all' will allow all dimensions to be preserved. In
+            this case, the result will be in the same shape/dimensionality
+            as the forecast, and the errors will be the error at each
+            point (i.e. single-value comparison against observed), and the
+            forecast and observed dimensions must match precisely.
+        scaling_factors : A 3-element vector or list describing the weights for each term in the KGE.
+            defined by: scaling_factors = [s_rho, s_alpha, s_beta] to apply to the correlation (rho),
+            variability (alpha), and bias (beta) terms respectively. Defaults to (1.0, 1.0, 1.0). (see
+            equation 10 in Gupta et al. (2009) for definitions of s_rho, s_alpha and s_beta)
[…]
