+++ tests/continuous/test_standard.py
@@ -689,6 +689,80 @@ EXP_PBIAS4 = xr.DataArray(np.array(-13 / 15.5 * 100))
 
 EXP_DS_PBIAS1 = xr.Dataset({"a": EXP_PBIAS1, "b": EXP_PBIAS2})
 
+## for KGE
+DA1_KGE = xr.DataArray(
+    np.array([[1, 2, 3], [0, 1, 0], [0.5, -0.5, 0.5], [3, 6, 3]]),
+    dims=("space", "time"),
+    coords=[
+        ("space", ["w", "x", "y", "z"]),
+        ("time", [1, 2, 3]),
+    ],
+)
+
+DA2_KGE = xr.DataArray(
+    np.array([[2, 4, 6], [6, 5, 6], [3, 4, 5], [3, np.nan, 3]]),
+    dims=("space", "time"),
+    coords=[
+        ("space", ["w", "x", "y", "z"]),
+        ("time", [1, 2, 3]),
+    ],
+)
+
+DA3_KGE = xr.DataArray(
+    np.array([[1, 2, 3], [3, 2.5, 3], [1.5, 2, 2.5], [1.5, np.nan, 1.5]]),
+    dims=("space", "time"),
+    coords=[
+        ("space", ["w", "x", "y", "z"]),
+        ("time", [1, 2, 3]),
+    ],
+)
+DA4_KGE = xr.DataArray(
+    np.array([[1, 3, 7], [2, 2, 8], [3, 1, 7]]),
+    dims=("space", "time"),
+    coords=[
+        ("space", ["x", "y", "z"]),
+        ("time", [1, 2, 3]),
+    ],
+)
+DA5_KGE = xr.DataArray(
+    np.array([1, 2, 3]),
+    dims=("space"),
+    coords=[("space", ["x", "y", "z"])],
+)
+
+## Expected KGE values
+EXP_KGE_KEEP_SPACE_DIM = xr.DataArray(
+    np.array([0.2928932188134524, -1.2103875562418747, -0.44811448882050064, np.nan]),
+    dims=("space"),
+    coords=[("space", ["w", "x", "y", "z"])],
+)
+EXP_KGE_REDUCE_ALL = xr.DataArray(0.2928932188134524)
+
+EXP_KGE_rho_returns_components = xr.DataArray(1.0)
+EXP_KGE_alpha_returns_components = xr.DataArray(0.5)
+EXP_KGE_beta_returns_components = xr.DataArray(0.5)
+
+EXP_KGE_returns_components = xr.Dataset(
+    {
+        "kge": EXP_KGE_REDUCE_ALL,
+        "rho": EXP_KGE_rho_returns_components,
+        "alpha": EXP_KGE_alpha_returns_components,
+        "beta": EXP_KGE_beta_returns_components,
+    }
+)
+
+
+EXP_KGE_Scaling_Factors = xr.DataArray(
+    1 - np.sqrt((0.5 * (1 - 1)) ** 2 + (1.0 * (0.5 - 1)) ** 2 + (2 * (0.5 - 1)) ** 2)
+)
+
+
+EXP_KGE_DIFF_SIZE = xr.DataArray(
+    np.array([1.0, -1.0, -1.8791915368841288]),
+    dims=("time"),
+    coords=[("time", [1, 2, 3])],
+)
+
 
 @pytest.mark.parametrize(
     ("fcst", "obs", "reduce_dims", "preserve_dims", "weights", "expected"),
@@ -813,3 +887,49 @@ def test_pbias_dask():
     result = result.compute()
     assert isinstance(result.data, np.ndarray)
     xr.testing.assert_equal(result, EXP_PBIAS3)
+
+
+@pytest.mark.parametrize(
[…]
