### 2024-11-01 nci/scores#553 @tennlee
https://scores.readthedocs.io/en/latest/tutorials/Threshold_Weighted_CRPS_for_Ensembles.html provides an end-to-end example tutorial now, but we still need something a bit better signposted and aimed at new users.
