### 2024-12-05 #6889 @tennlee
Hi there @crvernon @danielskatz - I wanted to politely follow up on the errors in the Crossref metadata, which we raised previously on July 10 and September 8.

As of today (December 5), there are still errors in the Crossref metadata. We are experiencing some downstream issues, which we believe may be caused by the incorrect Crossref metadata.

I believe only JOSS (as the publisher) can update Crossref metadata. I don’t believe we, as authors, can update Crossref metadata.

I notice that on both July 10 and September 9, @danielskatz pinged @openjournals/dev . I haven’t seen any response from @openjournals/dev on this thread. Is it possible that @openjournals/dev didn’t get a notification about this matter?

Even if the underlying reason for the errant newlines cannot be found, perhaps JOSS could manually update the metadata with Crossref?

Alternatively, is there any way that this issue could please be escalated?
