### 2024-12-05 #6889 @xuanxu
I've escalated this, we have now an open issue to investigate the origin of the newlines.
