### 2024-12-06 #6889 @xuanxu
We have updated the code generating the Crossref metadata and this has been fixed now.

@crvernon: the command `@editorialbot reaccept` should regenerate the files without newlines and redeposit with Crossref.
