### 2024-12-06 #6889 @tennlee
@xuanxu Thanks so much to you and the team for this. At a quick reading, Crossref is now looking okay. I'll see if anything else downstream changes in the next day or two, but to my eyes this looks great so far.
