### 2025-07-15 #6889 @paulrougieux
Dear @tennlee , your package uses Xarray as well. Would you like to review my paper [Cobwood: Enhancing Forest Economics Model Reusability Through Labeled Panel Data Structures](https://github.com/openjournals/joss-reviews/issues/8587) ?
